faer::col::generic

Struct Col

#[repr(transparent)]
pub struct Col<Inner>(pub Inner);

generic Col wrapper

Tuple Fields
0: Inner
Implementations

impl<'a, T> Col<Mut<'a, T>>

pub fn from_mut(value: &'a mut T) -> Self

creates a column view over the given element

pub fn from_slice_mut(slice: &'a mut [T]) -> Self

creates a ColMut from slice views over the column vector data, the result has the same number of rows as the length of the input slice

impl<'a, T, Rows: Shape, RStride: Stride> Col<Mut<'a, T, Rows, RStride>>

pub const unsafe fn from_raw_parts_mut( ptr: *mut T, nrows: Rows, row_stride: RStride, ) -> Self

creates a ColMut from pointers to the column vector data, number of rows, and row stride

safety

this function has the same safety requirements as [MatMut::from_raw_parts(ptr, nrows, 1, row_stride, 0)]

pub fn as_ptr(&self) -> *const T

returns a pointer to the column data

pub fn nrows(&self) -> Rows

returns the number of rows of the column

pub fn ncols(&self) -> usize

returns the number of columns of the column (always 1)

pub fn shape(&self) -> (Rows, usize)

returns the number of rows and columns of the column

pub fn row_stride(&self) -> RStride

returns the row stride of the column, specified in number of elements, not in bytes

pub fn ptr_at(&self, row: IdxInc<Rows>) -> *const T

returns a raw pointer to the element at the given index

pub unsafe fn ptr_inbounds_at(&self, row: Idx<Rows>) -> *const T

returns a raw pointer to the element at the given index, assuming the provided index is within the column bounds

safety

the behavior is undefined if any of the following conditions are violated:

row < self.nrows()

pub fn split_at_row( self, row: IdxInc<Rows>, ) -> (ColRef<'a, T, usize, RStride>, ColRef<'a, T, usize, RStride>)

see ColRef::split_at_row

pub fn transpose(self) -> RowRef<'a, T, Rows, RStride>

see ColRef::transpose

pub fn conjugate(self) -> ColRef<'a, T::Conj, Rows, RStride>
where T: Conjugate,

see ColRef::conjugate

pub fn canonical(self) -> ColRef<'a, T::Canonical, Rows, RStride>
where T: Conjugate,

see ColRef::canonical

pub fn adjoint(self) -> RowRef<'a, T::Conj, Rows, RStride>
where T: Conjugate,

see ColRef::adjoint

pub fn get<RowRange>( self, row: RowRange, ) -> <ColRef<'a, T, Rows, RStride> as ColIndex<RowRange>>::Target
where ColRef<'a, T, Rows, RStride>: ColIndex<RowRange>,

see ColRef::get

pub unsafe fn get_unchecked<RowRange>( self, row: RowRange, ) -> <ColRef<'a, T, Rows, RStride> as ColIndex<RowRange>>::Target
where ColRef<'a, T, Rows, RStride>: ColIndex<RowRange>,

see ColRef::get_unchecked

pub fn reverse_rows(self) -> ColRef<'a, T, Rows, RStride::Rev>

see ColRef::reverse_rows

pub fn subrows<V: Shape>( self, row_start: IdxInc<Rows>, nrows: V, ) -> ColRef<'a, T, V, RStride>

see ColRef::subrows

pub fn as_row_shape<V: Shape>(self, nrows: V) -> ColRef<'a, T, V, RStride>

see ColRef::as_row_shape

pub fn as_dyn_rows(self) -> ColRef<'a, T, usize, RStride>

see ColRef::as_dyn_rows

pub fn as_dyn_stride(self) -> ColRef<'a, T, Rows, isize>

see ColRef::as_dyn_stride

pub fn iter( self, ) -> impl 'a + ExactSizeIterator + DoubleEndedIterator<Item = &'a T>
where Rows: 'a,

see ColRef::iter

pub fn par_iter(self) -> impl 'a + IndexedParallelIterator<Item = &'a T>
where T: Sync, Rows: 'a,

see ColRef::par_iter

pub fn par_partition( self, count: usize, ) -> impl 'a + IndexedParallelIterator<Item = ColRef<'a, T, usize, RStride>>
where T: Sync, Rows: 'a,

see ColRef::par_partition

pub fn try_as_col_major(self) -> Option<ColRef<'a, T, Rows, ContiguousFwd>>

see ColRef::try_as_col_major

pub fn try_as_col_major_mut(self) -> Option<ColMut<'a, T, Rows, ContiguousFwd>>

see ColRef::try_as_col_major

pub fn as_mat(self) -> MatRef<'a, T, Rows, usize, RStride, isize>

see ColRef::as_mat

pub fn as_mat_mut(self) -> MatMut<'a, T, Rows, usize, RStride, isize>

see ColRef::as_mat

pub fn as_diagonal(self) -> DiagRef<'a, T, Rows, RStride>

see ColRef::as_diagonal

impl<T, Rows: Shape, RStride: Stride, Inner: for<'short> ReborrowMut<'short, Target = Mut<'short, T, Rows, RStride>>> Col<Inner>

pub fn copy_from<RhsT: Conjugate<Canonical = T>>( &mut self, other: impl AsColRef<T = RhsT, Rows = Rows>, )
where T: ComplexField,

copies other into self

pub fn fill(&mut self, value: T)
where T: Clone,

fills all the elements of self with value

pub fn as_mut(&mut self) -> ColMut<'_, T, Rows, RStride>

returns a view over self

impl<'a, T, Rows: Shape, RStride: Stride> Col<Mut<'a, T, Rows, RStride>>

pub fn as_ptr_mut(&self) -> *mut T

see ColRef::as_ptr

pub fn ptr_at_mut(&self, row: IdxInc<Rows>) -> *mut T

see ColRef::ptr_at

pub unsafe fn ptr_inbounds_at_mut(&self, row: Idx<Rows>) -> *mut T

see ColRef::ptr_inbounds_at

pub fn split_at_row_mut( self, row: IdxInc<Rows>, ) -> (ColMut<'a, T, usize, RStride>, ColMut<'a, T, usize, RStride>)

see ColRef::split_at_row

pub fn transpose_mut(self) -> RowMut<'a, T, Rows, RStride>

see ColRef::transpose

pub fn conjugate_mut(self) -> ColMut<'a, T::Conj, Rows, RStride>
where T: Conjugate,

see ColRef::conjugate

pub fn canonical_mut(self) -> ColMut<'a, T::Canonical, Rows, RStride>
where T: Conjugate,

see ColRef::canonical

pub fn adjoint_mut(self) -> RowMut<'a, T::Conj, Rows, RStride>
where T: Conjugate,

see ColRef::adjoint

pub fn get_mut<RowRange>( self, row: RowRange, ) -> <ColMut<'a, T, Rows, RStride> as ColIndex<RowRange>>::Target
where ColMut<'a, T, Rows, RStride>: ColIndex<RowRange>,

see ColRef::get

pub unsafe fn get_mut_unchecked<RowRange>( self, row: RowRange, ) -> <ColMut<'a, T, Rows, RStride> as ColIndex<RowRange>>::Target
where ColMut<'a, T, Rows, RStride>: ColIndex<RowRange>,

see ColRef::get_unchecked

pub fn reverse_rows_mut(self) -> ColMut<'a, T, Rows, RStride::Rev>

see ColRef::reverse_rows

pub fn subrows_mut<V: Shape>( self, row_start: IdxInc<Rows>, nrows: V, ) -> ColMut<'a, T, V, RStride>

see ColRef::subrows

pub fn as_row_shape_mut<V: Shape>(self, nrows: V) -> ColMut<'a, T, V, RStride>

see ColRef::as_row_shape

pub fn as_dyn_rows_mut(self) -> ColMut<'a, T, usize, RStride>

see ColRef::as_dyn_rows

pub fn as_dyn_stride_mut(self) -> ColMut<'a, T, Rows, isize>

see ColRef::as_dyn_stride

pub fn iter_mut( self, ) -> impl 'a + ExactSizeIterator + DoubleEndedIterator<Item = &'a mut T>
where Rows: 'a,

see ColRef::iter

pub fn par_iter_mut(self) -> impl 'a + IndexedParallelIterator<Item = &'a mut T>
where T: Send, Rows: 'a,

see ColRef::par_iter

pub fn par_partition_mut( self, count: usize, ) -> impl 'a + IndexedParallelIterator<Item = ColMut<'a, T, usize, RStride>>
where T: Send, Rows: 'a,

see ColRef::par_partition

pub fn as_diagonal_mut(self) -> DiagMut<'a, T, Rows, RStride>

see ColRef::as_diagonal

impl<'a, T, Rows: Shape> Col<Mut<'a, T, Rows, ContiguousFwd>>

pub fn as_slice_mut(self) -> &'a mut [T]

returns a reference over the elements as a slice

impl<'a, 'ROWS, T> Col<Mut<'a, T, Dim<'ROWS>, ContiguousFwd>>

pub fn as_array_mut(self) -> &'a mut Array<'ROWS, T>

returns a reference over the elements as a lifetime-bound slice

impl<'a, T, Rows: Shape> Col<Mut<'a, T, Rows>>
where T: RealField,

pub fn max(&self) -> Option<T>

Returns the maximum element in the column, or None if the column is empty

pub fn min(&self) -> Option<T>

Returns the minimum element in the column, or None if the column is empty

impl<T, Rows: Shape> Col<Own<T, Rows>>

pub fn from_fn(nrows: Rows, f: impl FnMut(Idx<Rows>) -> T) -> Self

returns a new column with dimension nrows, filled with the provided function

pub fn zeros(nrows: Rows) -> Self
where T: ComplexField,

returns a new column with dimension nrows, filled with zeros

pub fn ones(nrows: Rows) -> Self
where T: ComplexField,

returns a new column with dimension nrows, filled with ones

pub fn full(nrows: Rows, value: T) -> Self
where T: Clone,

returns a new column with dimension nrows, filled with value

pub fn try_reserve( &mut self, new_row_capacity: usize, ) -> Result<(), TryReserveError>

reserves the minimum capacity for row_capacity rows without reallocating, or returns an error in case of failure. does nothing if the capacity is already sufficient

pub fn reserve(&mut self, new_row_capacity: usize)

reserves the minimum capacity for row_capacity rows without reallocating. does nothing if the capacity is already sufficient

pub fn resize_with(&mut self, new_nrows: Rows, f: impl FnMut(Idx<Rows>) -> T)

resizes the column in-place so that the new dimension is new_nrows. new elements are created with the given function f, so that elements at index i are created by calling f(i)

pub fn truncate(&mut self, new_nrows: Rows)

truncates the column so that its new dimensions are new_nrows.
the new dimension must be smaller than or equal to the current dimension

panics

the function panics if any of the following conditions are violated:

new_nrows > self.nrows()

pub fn into_row_shape<V: Shape>(self, nrows: V) -> Col<T, V>

see ColRef::as_row_shape

pub fn into_diagonal(self) -> Diag<T, Rows>

see ColRef::as_diagonal

pub fn into_transpose(self) -> Row<T, Rows>

see ColRef::transpose

impl<T, Rows: Shape> Col<Own<T, Rows>>

pub fn nrows(&self) -> Rows

returns the number of rows of the column

pub fn ncols(&self) -> usize

returns the number of columns of the column (always 1)

impl<T, Rows: Shape> Col<Own<T, Rows>>

pub fn as_ptr(&self) -> *const T

see ColRef::as_ptr

pub fn shape(&self) -> (Rows, usize)

see ColRef::shape

pub fn row_stride(&self) -> isize

see ColRef::row_stride

pub fn ptr_at(&self, row: IdxInc<Rows>) -> *const T

see ColRef::ptr_at

pub unsafe fn ptr_inbounds_at(&self, row: Idx<Rows>) -> *const T

see ColRef::ptr_inbounds_at

pub fn split_at_row( &self, row: IdxInc<Rows>, ) -> (ColRef<'_, T, usize>, ColRef<'_, T, usize>)

see ColRef::split_at_row

pub fn transpose(&self) -> RowRef<'_, T, Rows>

see ColRef::transpose

pub fn conjugate(&self) -> ColRef<'_, T::Conj, Rows>
where T: Conjugate,

see ColRef::conjugate

pub fn canonical(&self) -> ColRef<'_, T::Canonical, Rows>
where T: Conjugate,

see ColRef::canonical

pub fn adjoint(&self) -> RowRef<'_, T::Conj, Rows>
where T: Conjugate,

see ColRef::adjoint

pub fn get<RowRange>( &self, row: RowRange, ) -> <ColRef<'_, T, Rows> as ColIndex<RowRange>>::Target
where for<'a> ColRef<'a, T, Rows>: ColIndex<RowRange>,

see ColRef::get

pub unsafe fn get_unchecked<RowRange>( &self, row: RowRange, ) -> <ColRef<'_, T, Rows> as ColIndex<RowRange>>::Target
where for<'a> ColRef<'a, T, Rows>: ColIndex<RowRange>,

see ColRef::get_unchecked

pub fn reverse_rows(&self) -> ColRef<'_, T, Rows>

see ColRef::reverse_rows

pub fn subrows<V: Shape>( &self, row_start: IdxInc<Rows>, nrows: V, ) -> ColRef<'_, T, V>

see ColRef::subrows

pub fn as_row_shape<V: Shape>(&self, nrows: V) -> ColRef<'_, T, V>

see ColRef::as_row_shape

pub fn as_dyn_rows(&self) -> ColRef<'_, T, usize>

see ColRef::as_dyn_rows

pub fn as_dyn_stride(&self) -> ColRef<'_, T, Rows, isize>

see ColRef::as_dyn_stride

pub fn iter( &self, ) -> impl '_ + ExactSizeIterator + DoubleEndedIterator<Item = &T>

see ColRef::iter

pub fn par_iter(&self) -> impl '_ + IndexedParallelIterator<Item = &T>
where T: Sync,

see ColRef::par_iter

pub fn par_partition( &self, count: usize, ) -> impl '_ + IndexedParallelIterator<Item = ColRef<'_, T, usize>>
where T: Sync,

see ColRef::par_partition

pub fn try_as_col_major(&self) -> Option<ColRef<'_, T, Rows, ContiguousFwd>>

see ColRef::try_as_col_major

pub fn try_as_col_major_mut( &mut self, ) -> Option<ColMut<'_, T, Rows, ContiguousFwd>>

see ColRef::try_as_col_major

pub fn as_mat(&self) -> MatRef<'_, T, Rows, usize, isize>

see ColRef::as_mat

pub fn as_mat_mut(&mut self) -> MatMut<'_, T, Rows, usize, isize>

see ColRef::as_mat

pub fn as_diagonal(&self) -> DiagRef<'_, T, Rows>

see ColRef::as_diagonal

impl<T, Rows: Shape> Col<Own<T, Rows>>

pub fn as_ptr_mut(&mut self) -> *mut T

see ColMut::as_ptr_mut

pub fn ptr_at_mut(&mut self, row: IdxInc<Rows>) -> *mut T

see ColMut::ptr_at_mut

pub unsafe fn ptr_inbounds_at_mut(&mut self, row: Idx<Rows>) -> *mut T

see ColMut::ptr_inbounds_at_mut

pub fn split_at_row_mut( &mut self, row: IdxInc<Rows>, ) -> (ColMut<'_, T, usize>, ColMut<'_, T, usize>)

see ColMut::split_at_row_mut

pub fn transpose_mut(&mut self) -> RowMut<'_, T, Rows>

see ColMut::transpose_mut

pub fn conjugate_mut(&mut self) -> ColMut<'_, T::Conj, Rows>
where T: Conjugate,

see ColMut::conjugate_mut

pub fn canonical_mut(&mut self) -> ColMut<'_, T::Canonical, Rows>
where T: Conjugate,

see ColMut::canonical_mut

pub fn adjoint_mut(&mut self) -> RowMut<'_, T::Conj, Rows>
where T: Conjugate,

see ColMut::adjoint_mut

pub fn get_mut<RowRange>( &mut self, row: RowRange, ) -> <ColMut<'_, T, Rows> as ColIndex<RowRange>>::Target
where for<'a> ColMut<'a, T, Rows>: ColIndex<RowRange>,

see ColMut::get_mut

pub unsafe fn get_mut_unchecked<RowRange>( &mut self, row: RowRange, ) -> <ColMut<'_, T, Rows> as ColIndex<RowRange>>::Target
where for<'a> ColMut<'a, T, Rows>: ColIndex<RowRange>,

see ColMut::get_mut_unchecked

pub fn reverse_rows_mut(&mut self) -> ColMut<'_, T, Rows>

see ColMut::reverse_rows_mut

pub fn subrows_mut<V: Shape>( &mut self, row_start: IdxInc<Rows>, nrows: V, ) -> ColMut<'_, T, V>

see ColMut::subrows_mut

pub fn as_row_shape_mut<V: Shape>(&mut self, nrows: V) -> ColMut<'_, T, V>

see ColMut::as_row_shape_mut

pub fn as_dyn_rows_mut(&mut self) -> ColMut<'_, T, usize>

see ColMut::as_dyn_rows_mut

pub fn as_dyn_stride_mut(&mut self) -> ColMut<'_, T, Rows, isize>

see ColMut::as_dyn_stride_mut

pub fn iter_mut( &mut self, ) -> impl '_ + ExactSizeIterator + DoubleEndedIterator<Item = &mut T>

see ColMut::iter_mut

pub fn par_iter_mut( &mut self, ) -> impl '_ + IndexedParallelIterator<Item = &mut T>
where T: Send,

see ColMut::par_iter_mut

pub fn par_partition_mut( &mut self, count: usize, ) -> impl '_ + IndexedParallelIterator<Item = ColMut<'_, T, usize>>
where T: Send,

see ColMut::par_partition_mut

pub fn as_diagonal_mut(&mut self) -> DiagMut<'_, T, Rows>

see ColMut::as_diagonal_mut

impl<T, Rows: Shape> Col<Own<T, Rows>>
where T: RealField,

pub fn max(&self) -> Option<T>

Returns the maximum element in the column, or None if the column is empty

pub fn min(&self) -> Option<T>

Returns the minimum element in the column, or None if the column is empty

impl<'a, T> Col<Ref<'a, T>>

pub fn from_ref(value: &'a T) -> Self

creates a column view over the given element

pub fn from_slice(slice: &'a [T]) -> Self

creates a ColRef from slice views over the column vector data, the result has the same number of rows as the length of the input slice

impl<'a, T, Rows: Shape, RStride: Stride> Col<Ref<'a, T, Rows, RStride>>

pub const unsafe fn from_raw_parts( ptr: *const T, nrows: Rows, row_stride: RStride, ) -> Self

creates a ColRef from pointers to the column vector data, number of rows, and row stride

safety

this function has the same safety requirements as [MatRef::from_raw_parts(ptr, nrows, 1, row_stride, 0)]

pub fn as_ptr(&self) -> *const T

returns a pointer to the column data

pub fn nrows(&self) -> Rows

returns the number of rows of the column

pub fn ncols(&self) -> usize

returns the number of columns of the column (always 1)

pub fn shape(&self) -> (Rows, usize)

returns the number of rows and columns of the column

pub fn row_stride(&self) -> RStride

returns the row stride of the column, specified in number of elements, not in bytes

pub fn ptr_at(&self, row: IdxInc<Rows>) -> *const T

returns a raw pointer to the element at the given index

pub unsafe fn ptr_inbounds_at(&self, row: Idx<Rows>) -> *const T

returns a raw pointer to the element at the given index, assuming the provided index is within the column bounds

safety

the behavior is undefined if any of the following conditions are violated:

row < self.nrows()

pub fn split_at_row( self, row: IdxInc<Rows>, ) -> (ColRef<'a, T, usize, RStride>, ColRef<'a, T, usize, RStride>)

splits the column horizontally at the given row into two parts and returns an array of each submatrix, in the following order:

top

bottom

panics

the function panics if the following condition is violated:

row <= self.nrows()

pub fn transpose(self) -> RowRef<'a, T, Rows, RStride>

returns a view over the transpose of self

pub fn conjugate(self) -> ColRef<'a, T::Conj, Rows, RStride>
where T: Conjugate,

returns a view over the conjugate of self

pub fn canonical(self) -> ColRef<'a, T::Canonical, Rows, RStride>
where T: Conjugate,

returns an unconjugated view over self

pub fn adjoint(self) -> RowRef<'a, T::Conj, Rows, RStride>
where T: Conjugate,

returns a view over the conjugate transpose of self

pub fn get<RowRange>( self, row: RowRange, ) -> <ColRef<'a, T, Rows, RStride> as ColIndex<RowRange>>::Target
where ColRef<'a, T, Rows, RStride>: ColIndex<RowRange>,

returns a reference to the element at the given index, or a subcolumn if row is a range

panics

the function panics if any of the following conditions are violated:

row must be contained in [0, self.nrows())

pub unsafe fn get_unchecked<RowRange>( self, row: RowRange, ) -> <ColRef<'a, T, Rows, RStride> as ColIndex<RowRange>>::Target
where ColRef<'a, T, Rows, RStride>: ColIndex<RowRange>,

returns a reference to the element at the given index, or a subcolumn if row is a range, without bound checks

safety

the behavior is undefined if any of the following conditions are violated:

row must be contained in [0, self.nrows())

pub fn reverse_rows(self) -> ColRef<'a, T, Rows, RStride::Rev>

returns a view over the self, with the rows in reversed order

pub fn subrows<V: Shape>( self, row_start: IdxInc<Rows>, nrows: V, ) -> ColRef<'a, T, V, RStride>

returns a view over the column starting at row row_start, and with number of rows nrows

panics

the function panics if any of the following conditions are violated:

row_start <= self.nrows()

nrows <= self.nrows() - row_start

pub fn as_row_shape<V: Shape>(self, nrows: V) -> ColRef<'a, T, V, RStride>

returns the input column with the given row shape after checking that it matches the current row shape

pub fn as_dyn_rows(self) -> ColRef<'a, T, usize, RStride>

returns the input column with dynamic row shape

pub fn as_dyn_stride(self) -> ColRef<'a, T, Rows, isize>

returns the input column with dynamic stride

pub fn iter( self, ) -> impl 'a + ExactSizeIterator + DoubleEndedIterator<Item = &'a T>
where Rows: 'a,

returns an iterator over the elements of the column

pub fn par_iter(self) -> impl 'a + IndexedParallelIterator<Item = &'a T>
where T: Sync, Rows: 'a,

returns a parallel iterator over the elements of the column

pub fn par_partition( self, count: usize, ) -> impl 'a + IndexedParallelIterator<Item = ColRef<'a, T, usize, RStride>>
where T: Sync, Rows: 'a,

returns a parallel iterator that provides exactly count successive chunks of the elements of this column

only available with the rayon feature

pub fn try_as_col_major(self) -> Option<ColRef<'a, T, Rows, ContiguousFwd>>

returns a view over the column with a static row stride equal to +1, or None otherwise

pub fn as_mat(self) -> MatRef<'a, T, Rows, usize, RStride, isize>

returns a matrix view over self

pub fn as_diagonal(self) -> DiagRef<'a, T, Rows, RStride>

interprets the column as a diagonal matrix

impl<T, Rows: Shape, RStride: Stride, Inner: for<'short> Reborrow<'short, Target = Ref<'short, T, Rows, RStride>>> Col<Inner>

pub fn cloned(&self) -> Col<T, Rows>
where T: Clone,

returns a newly allocated column holding the cloned values of self

pub fn to_owned(&self) -> Col<T::Canonical, Rows>
where T: Conjugate,

returns a newly allocated column holding the (possibly conjugated) values of self

pub fn norm_max(&self) -> Real<T>
where T: Conjugate,

returns the maximum norm of self

pub fn norm_l2(&self) -> Real<T>
where T: Conjugate,

returns the l2 norm of self

pub fn squared_norm_l2(&self) -> Real<T>
where T: Conjugate,

returns the squared l2 norm of self

pub fn norm_l1(&self) -> Real<T>
where T: Conjugate,

returns the l1 norm of self

pub fn sum(&self) -> T::Canonical
where T: Conjugate,

returns the sum of the elements of self

pub fn as_ref(&self) -> ColRef<'_, T, Rows, RStride>

returns a view over self

pub fn kron( &self, rhs: impl AsMatRef<T: Conjugate<Canonical = T::Canonical>>, ) -> Mat<T::Canonical>
where T: Conjugate,

see Mat::kron

pub fn is_all_finite(&self) -> bool
where T: Conjugate,

returns true if all of the elements of self are finite. otherwise returns false.

pub fn has_nan(&self) -> bool
where T: Conjugate,

returns true if any of the elements of self is NaN. otherwise returns false.

impl<'a, T, Rows: Shape> Col<Ref<'a, T, Rows, ContiguousFwd>>

pub fn as_slice(self) -> &'a [T]

returns a reference over the elements as a slice

impl<'a, 'ROWS, T> Col<Ref<'a, T, Dim<'ROWS>, ContiguousFwd>>

pub fn as_array(self) -> &'a Array<'ROWS, T>

returns a reference over the elements as a lifetime-bound slice

impl<'a, T, Rows: Shape, RStride: Stride> Col<Ref<'a, T, Rows, RStride>>
where T: RealField,

pub fn max(&self) -> Option<T>

Returns the maximum element in the column, or None if the column is empty

pub fn min(&self) -> Option<T>

Returns the minimum element in the column, or None if the column is empty

impl<Inner> Col<Inner>

pub fn from_inner_ref(inner: &Inner) -> &Self

wrap by reference

pub fn from_inner_mut(inner: &mut Inner) -> &mut Self

wrap by mutable reference