faer::col

Type Alias Col

pub type Col<T, Rows = usize> = Col<Own<T, Rows>>;

heap allocated resizable column vector.

note

the memory layout of Col is guaranteed to be column-major, meaning that it has a row stride of 1.

Aliased Type

#[repr(transparent)]
pub struct Col<T, Rows = usize>(pub Own<T, Rows>);

Tuple Fields
0: Own<T, Rows>
Implementations

impl<T, Rows: Shape> Col<T, Rows>

pub fn from_fn(nrows: Rows, f: impl FnMut(Idx<Rows>) -> T) -> Self

returns a new column with dimension nrows, filled with the provided function

pub fn zeros(nrows: Rows) -> Self
where T: ComplexField,

returns a new column with dimension nrows, filled with zeros

pub fn ones(nrows: Rows) -> Self
where T: ComplexField,

returns a new column with dimension nrows, filled with ones

pub fn full(nrows: Rows, value: T) -> Self
where T: Clone,

returns a new column with dimension nrows, filled with value

pub fn try_reserve( &mut self, new_row_capacity: usize, ) -> Result<(), TryReserveError>

reserves the minimum capacity for row_capacity rows without reallocating, or returns an error in case of failure. does nothing if the capacity is already sufficient

pub fn reserve(&mut self, new_row_capacity: usize)

reserves the minimum capacity for row_capacity rows without reallocating. does nothing if the capacity is already sufficient

pub fn resize_with(&mut self, new_nrows: Rows, f: impl FnMut(Idx<Rows>) -> T)

resizes the column in-place so that the new dimension is new_nrows. new elements are created with the given function f, so that elements at index i are created by calling f(i)

pub fn truncate(&mut self, new_nrows: Rows)

truncates the column so that its new dimensions are new_nrows.
the new dimension must be smaller than or equal to the current dimension

panics

the function panics if any of the following conditions are violated:

new_nrows > self.nrows()

pub fn into_row_shape<V: Shape>(self, nrows: V) -> Col<T, V>

see ColRef::as_row_shape

pub fn into_diagonal(self) -> Diag<T, Rows>

see ColRef::as_diagonal

pub fn into_transpose(self) -> Row<T, Rows>

see ColRef::transpose

impl<T, Rows: Shape> Col<T, Rows>

pub fn nrows(&self) -> Rows

returns the number of rows of the column

pub fn ncols(&self) -> usize

returns the number of columns of the column (always 1)

impl<T, Rows: Shape> Col<T, Rows>

pub fn as_ptr(&self) -> *const T

see ColRef::as_ptr

pub fn shape(&self) -> (Rows, usize)

see ColRef::shape

pub fn row_stride(&self) -> isize

see ColRef::row_stride

pub fn ptr_at(&self, row: IdxInc<Rows>) -> *const T

see ColRef::ptr_at

pub unsafe fn ptr_inbounds_at(&self, row: Idx<Rows>) -> *const T

see ColRef::ptr_inbounds_at

pub fn split_at_row( &self, row: IdxInc<Rows>, ) -> (ColRef<'_, T, usize>, ColRef<'_, T, usize>)

see ColRef::split_at_row

pub fn transpose(&self) -> RowRef<'_, T, Rows>

see ColRef::transpose

pub fn conjugate(&self) -> ColRef<'_, T::Conj, Rows>
where T: Conjugate,

see ColRef::conjugate

pub fn canonical(&self) -> ColRef<'_, T::Canonical, Rows>
where T: Conjugate,

see ColRef::canonical

pub fn adjoint(&self) -> RowRef<'_, T::Conj, Rows>
where T: Conjugate,

see ColRef::adjoint

pub fn get<RowRange>( &self, row: RowRange, ) -> <ColRef<'_, T, Rows> as ColIndex<RowRange>>::Target
where for<'a> ColRef<'a, T, Rows>: ColIndex<RowRange>,

see ColRef::get

pub unsafe fn get_unchecked<RowRange>( &self, row: RowRange, ) -> <ColRef<'_, T, Rows> as ColIndex<RowRange>>::Target
where for<'a> ColRef<'a, T, Rows>: ColIndex<RowRange>,

see ColRef::get_unchecked

pub fn reverse_rows(&self) -> ColRef<'_, T, Rows>

see ColRef::reverse_rows

pub fn subrows<V: Shape>( &self, row_start: IdxInc<Rows>, nrows: V, ) -> ColRef<'_, T, V>

see ColRef::subrows

pub fn as_row_shape<V: Shape>(&self, nrows: V) -> ColRef<'_, T, V>

see ColRef::as_row_shape

pub fn as_dyn_rows(&self) -> ColRef<'_, T, usize>

see ColRef::as_dyn_rows

pub fn as_dyn_stride(&self) -> ColRef<'_, T, Rows, isize>

see ColRef::as_dyn_stride

pub fn iter( &self, ) -> impl '_ + ExactSizeIterator + DoubleEndedIterator<Item = &T>

see ColRef::iter

pub fn par_iter(&self) -> impl '_ + IndexedParallelIterator<Item = &T>
where T: Sync,

see ColRef::par_iter

pub fn par_partition( &self, count: usize, ) -> impl '_ + IndexedParallelIterator<Item = ColRef<'_, T, usize>>
where T: Sync,

see ColRef::par_partition

pub fn try_as_col_major(&self) -> Option<ColRef<'_, T, Rows, ContiguousFwd>>

see ColRef::try_as_col_major

pub fn try_as_col_major_mut( &mut self, ) -> Option<ColMut<'_, T, Rows, ContiguousFwd>>

see ColRef::try_as_col_major

pub fn as_mat(&self) -> MatRef<'_, T, Rows, usize, isize>

see ColRef::as_mat

pub fn as_mat_mut(&mut self) -> MatMut<'_, T, Rows, usize, isize>

see ColRef::as_mat

pub fn as_diagonal(&self) -> DiagRef<'_, T, Rows>

see ColRef::as_diagonal

impl<T, Rows: Shape> Col<T, Rows>

pub fn as_ptr_mut(&mut self) -> *mut T

see ColMut::as_ptr_mut

pub fn ptr_at_mut(&mut self, row: IdxInc<Rows>) -> *mut T

see ColMut::ptr_at_mut

pub unsafe fn ptr_inbounds_at_mut(&mut self, row: Idx<Rows>) -> *mut T

see ColMut::ptr_inbounds_at_mut

pub fn split_at_row_mut( &mut self, row: IdxInc<Rows>, ) -> (ColMut<'_, T, usize>, ColMut<'_, T, usize>)

see ColMut::split_at_row_mut

pub fn transpose_mut(&mut self) -> RowMut<'_, T, Rows>

see ColMut::transpose_mut

pub fn conjugate_mut(&mut self) -> ColMut<'_, T::Conj, Rows>
where T: Conjugate,

see ColMut::conjugate_mut

pub fn canonical_mut(&mut self) -> ColMut<'_, T::Canonical, Rows>
where T: Conjugate,

see ColMut::canonical_mut

pub fn adjoint_mut(&mut self) -> RowMut<'_, T::Conj, Rows>
where T: Conjugate,

see ColMut::adjoint_mut

pub fn get_mut<RowRange>( &mut self, row: RowRange, ) -> <ColMut<'_, T, Rows> as ColIndex<RowRange>>::Target
where for<'a> ColMut<'a, T, Rows>: ColIndex<RowRange>,

see ColMut::get_mut

pub unsafe fn get_mut_unchecked<RowRange>( &mut self, row: RowRange, ) -> <ColMut<'_, T, Rows> as ColIndex<RowRange>>::Target
where for<'a> ColMut<'a, T, Rows>: ColIndex<RowRange>,

see ColMut::get_mut_unchecked

pub fn reverse_rows_mut(&mut self) -> ColMut<'_, T, Rows>

see ColMut::reverse_rows_mut

pub fn subrows_mut<V: Shape>( &mut self, row_start: IdxInc<Rows>, nrows: V, ) -> ColMut<'_, T, V>

see ColMut::subrows_mut

pub fn as_row_shape_mut<V: Shape>(&mut self, nrows: V) -> ColMut<'_, T, V>

see ColMut::as_row_shape_mut

pub fn as_dyn_rows_mut(&mut self) -> ColMut<'_, T, usize>

see ColMut::as_dyn_rows_mut

pub fn as_dyn_stride_mut(&mut self) -> ColMut<'_, T, Rows, isize>

see ColMut::as_dyn_stride_mut

pub fn iter_mut( &mut self, ) -> impl '_ + ExactSizeIterator + DoubleEndedIterator<Item = &mut T>

see ColMut::iter_mut

pub fn par_iter_mut( &mut self, ) -> impl '_ + IndexedParallelIterator<Item = &mut T>
where T: Send,

see ColMut::par_iter_mut

pub fn par_partition_mut( &mut self, count: usize, ) -> impl '_ + IndexedParallelIterator<Item = ColMut<'_, T, usize>>
where T: Send,

see ColMut::par_partition_mut

pub fn as_diagonal_mut(&mut self) -> DiagMut<'_, T, Rows>

see ColMut::as_diagonal_mut

impl<T, Rows: Shape> Col<T, Rows>
where T: RealField,

pub fn max(&self) -> Option<T>

Returns the maximum element in the column, or None if the column is empty

pub fn min(&self) -> Option<T>

Returns the minimum element in the column, or None if the column is empty