faer::col

Type Alias ColMut

pub type ColMut<'a, T, Rows = usize, RStride = isize> = Col<Mut<'a, T, Rows, RStride>>;

mutable view over a column vector, similar to a mutable reference to a strided slice

note

unlike a slice, the data pointed to by ColMut<'_, T> is allowed to be partially or fully uninitialized under certain conditions. in this case, care must be taken to not perform any operations that read the uninitialized values, or form references to them, either directly or indirectly through any of the numerical library routines, unless it is explicitly permitted

Aliased Type

struct ColMut<'a, T, Rows = usize, RStride = isize>(pub Mut<'a, T, Rows, RStride>);

Fields
0: Mut<'a, T, Rows, RStride>
Implementations

impl<'a, T> ColMut<'a, T>

pub fn from_mut(value: &'a mut T) -> Self

creates a column view over the given element

pub fn from_slice_mut(slice: &'a mut [T]) -> Self

creates a ColMut from slice views over the column vector data, the result has the same number of rows as the length of the input slice

impl<'a, T, Rows: Shape, RStride: Stride> ColMut<'a, T, Rows, RStride>

pub const unsafe fn from_raw_parts_mut( ptr: *mut T, nrows: Rows, row_stride: RStride, ) -> Self

creates a ColMut from pointers to the column vector data, number of rows, and row stride

safety

this function has the same safety requirements as [MatMut::from_raw_parts(ptr, nrows, 1, row_stride, 0)]

pub fn as_ptr(&self) -> *const T

returns a pointer to the column data

pub fn nrows(&self) -> Rows

returns the number of rows of the column

pub fn ncols(&self) -> usize

returns the number of columns of the column (always 1)

pub fn shape(&self) -> (Rows, usize)

returns the number of rows and columns of the column

pub fn row_stride(&self) -> RStride

returns the row stride of the column, specified in number of elements, not in bytes

pub fn ptr_at(&self, row: IdxInc<Rows>) -> *const T

returns a raw pointer to the element at the given index

pub unsafe fn ptr_inbounds_at(&self, row: Idx<Rows>) -> *const T

returns a raw pointer to the element at the given index, assuming the provided index is within the column bounds

safety

the behavior is undefined if any of the following conditions are violated:

row < self.nrows()

pub fn split_at_row( self, row: IdxInc<Rows>, ) -> (ColRef<'a, T, usize, RStride>, ColRef<'a, T, usize, RStride>)

see ColRef::split_at_row

pub fn transpose(self) -> RowRef<'a, T, Rows, RStride>

see ColRef::transpose

pub fn conjugate(self) -> ColRef<'a, T::Conj, Rows, RStride>
where T: Conjugate,

see ColRef::conjugate

pub fn canonical(self) -> ColRef<'a, T::Canonical, Rows, RStride>
where T: Conjugate,

see ColRef::canonical

pub fn adjoint(self) -> RowRef<'a, T::Conj, Rows, RStride>
where T: Conjugate,

see ColRef::adjoint

pub fn get<RowRange>( self, row: RowRange, ) -> <ColRef<'a, T, Rows, RStride> as ColIndex<RowRange>>::Target
where ColRef<'a, T, Rows, RStride>: ColIndex<RowRange>,

see ColRef::get

pub unsafe fn get_unchecked<RowRange>( self, row: RowRange, ) -> <ColRef<'a, T, Rows, RStride> as ColIndex<RowRange>>::Target
where ColRef<'a, T, Rows, RStride>: ColIndex<RowRange>,

see ColRef::get_unchecked

pub fn reverse_rows(self) -> ColRef<'a, T, Rows, RStride::Rev>

see ColRef::reverse_rows

pub fn subrows<V: Shape>( self, row_start: IdxInc<Rows>, nrows: V, ) -> ColRef<'a, T, V, RStride>

see ColRef::subrows

pub fn as_row_shape<V: Shape>(self, nrows: V) -> ColRef<'a, T, V, RStride>

see ColRef::as_row_shape

pub fn as_dyn_rows(self) -> ColRef<'a, T, usize, RStride>

see ColRef::as_dyn_rows

pub fn as_dyn_stride(self) -> ColRef<'a, T, Rows, isize>

see ColRef::as_dyn_stride

pub fn iter( self, ) -> impl 'a + ExactSizeIterator + DoubleEndedIterator<Item = &'a T>
where Rows: 'a,

see ColRef::iter

pub fn par_iter(self) -> impl 'a + IndexedParallelIterator<Item = &'a T>
where T: Sync, Rows: 'a,

see ColRef::par_iter

pub fn par_partition( self, count: usize, ) -> impl 'a + IndexedParallelIterator<Item = ColRef<'a, T, usize, RStride>>
where T: Sync, Rows: 'a,

see ColRef::par_partition

pub fn try_as_col_major(self) -> Option<ColRef<'a, T, Rows, ContiguousFwd>>

see ColRef::try_as_col_major

pub fn try_as_col_major_mut(self) -> Option<ColMut<'a, T, Rows, ContiguousFwd>>

see ColRef::try_as_col_major

pub fn as_mat(self) -> MatRef<'a, T, Rows, usize, RStride, isize>

see ColRef::as_mat

pub fn as_mat_mut(self) -> MatMut<'a, T, Rows, usize, RStride, isize>

see ColRef::as_mat

pub fn as_diagonal(self) -> DiagRef<'a, T, Rows, RStride>

see ColRef::as_diagonal

impl<'a, T, Rows: Shape, RStride: Stride> ColMut<'a, T, Rows, RStride>

pub fn as_ptr_mut(&self) -> *mut T

see ColRef::as_ptr

pub fn ptr_at_mut(&self, row: IdxInc<Rows>) -> *mut T

see ColRef::ptr_at

pub unsafe fn ptr_inbounds_at_mut(&self, row: Idx<Rows>) -> *mut T

see ColRef::ptr_inbounds_at

pub fn split_at_row_mut( self, row: IdxInc<Rows>, ) -> (ColMut<'a, T, usize, RStride>, ColMut<'a, T, usize, RStride>)

see ColRef::split_at_row

pub fn transpose_mut(self) -> RowMut<'a, T, Rows, RStride>

see ColRef::transpose

pub fn conjugate_mut(self) -> ColMut<'a, T::Conj, Rows, RStride>
where T: Conjugate,

see ColRef::conjugate

pub fn canonical_mut(self) -> ColMut<'a, T::Canonical, Rows, RStride>
where T: Conjugate,

see ColRef::canonical

pub fn adjoint_mut(self) -> RowMut<'a, T::Conj, Rows, RStride>
where T: Conjugate,

see ColRef::adjoint

pub fn get_mut<RowRange>( self, row: RowRange, ) -> <ColMut<'a, T, Rows, RStride> as ColIndex<RowRange>>::Target
where ColMut<'a, T, Rows, RStride>: ColIndex<RowRange>,

see ColRef::get

pub unsafe fn get_mut_unchecked<RowRange>( self, row: RowRange, ) -> <ColMut<'a, T, Rows, RStride> as ColIndex<RowRange>>::Target
where ColMut<'a, T, Rows, RStride>: ColIndex<RowRange>,

see ColRef::get_unchecked

pub fn reverse_rows_mut(self) -> ColMut<'a, T, Rows, RStride::Rev>

see ColRef::reverse_rows

pub fn subrows_mut<V: Shape>( self, row_start: IdxInc<Rows>, nrows: V, ) -> ColMut<'a, T, V, RStride>

see ColRef::subrows

pub fn as_row_shape_mut<V: Shape>(self, nrows: V) -> ColMut<'a, T, V, RStride>

see ColRef::as_row_shape

pub fn as_dyn_rows_mut(self) -> ColMut<'a, T, usize, RStride>

see ColRef::as_dyn_rows

pub fn as_dyn_stride_mut(self) -> ColMut<'a, T, Rows, isize>

see ColRef::as_dyn_stride

pub fn iter_mut( self, ) -> impl 'a + ExactSizeIterator + DoubleEndedIterator<Item = &'a mut T>
where Rows: 'a,

see ColRef::iter

pub fn par_iter_mut(self) -> impl 'a + IndexedParallelIterator<Item = &'a mut T>
where T: Send, Rows: 'a,

see ColRef::par_iter

pub fn par_partition_mut( self, count: usize, ) -> impl 'a + IndexedParallelIterator<Item = ColMut<'a, T, usize, RStride>>
where T: Send, Rows: 'a,

see ColRef::par_partition

pub fn as_diagonal_mut(self) -> DiagMut<'a, T, Rows, RStride>

see ColRef::as_diagonal

impl<'a, T, Rows: Shape> ColMut<'a, T, Rows, ContiguousFwd>

pub fn as_slice_mut(self) -> &'a mut [T]

returns a reference over the elements as a slice

impl<'a, 'ROWS, T> ColMut<'a, T, Dim<'ROWS>, ContiguousFwd>

pub fn as_array_mut(self) -> &'a mut Array<'ROWS, T>

returns a reference over the elements as a lifetime-bound slice

impl<'a, T, Rows: Shape> ColMut<'a, T, Rows>
where T: RealField,

pub fn max(&self) -> Option<T>

Returns the maximum element in the column, or None if the column is empty

pub fn min(&self) -> Option<T>

Returns the minimum element in the column, or None if the column is empty