faer::col

Type Alias ColRef

pub type ColRef<'a, T, Rows = usize, RStride = isize> = Col<Ref<'a, T, Rows, RStride>>;

immutable view over a column vector, similar to an immutable reference to a strided slice

note

unlike a slice, the data pointed to by ColRef<'_, T> is allowed to be partially or fully uninitialized under certain conditions. in this case, care must be taken to not perform any operations that read the uninitialized values, or form references to them, either directly or indirectly through any of the numerical library routines, unless it is explicitly permitted

Aliased Type

#[repr(transparent)]
pub struct ColRef<'a, T, Rows = usize, RStride = isize>(pub Ref<'a, T, Rows, RStride>);

Tuple Fields
0: Ref<'a, T, Rows, RStride>
Implementations

impl<'a, T> ColRef<'a, T>

pub fn from_ref(value: &'a T) -> Self

creates a column view over the given element

pub fn from_slice(slice: &'a [T]) -> Self

creates a ColRef from slice views over the column vector data, the result has the same number of rows as the length of the input slice

impl<'a, T, Rows: Shape, RStride: Stride> ColRef<'a, T, Rows, RStride>

pub const unsafe fn from_raw_parts( ptr: *const T, nrows: Rows, row_stride: RStride, ) -> Self

creates a ColRef from pointers to the column vector data, number of rows, and row stride

safety

this function has the same safety requirements as [MatRef::from_raw_parts(ptr, nrows, 1, row_stride, 0)]

pub fn as_ptr(&self) -> *const T

returns a pointer to the column data

pub fn nrows(&self) -> Rows

returns the number of rows of the column

pub fn ncols(&self) -> usize

returns the number of columns of the column (always 1)

pub fn shape(&self) -> (Rows, usize)

returns the number of rows and columns of the column

pub fn row_stride(&self) -> RStride

returns the row stride of the column, specified in number of elements, not in bytes

pub fn ptr_at(&self, row: IdxInc<Rows>) -> *const T

returns a raw pointer to the element at the given index

pub unsafe fn ptr_inbounds_at(&self, row: Idx<Rows>) -> *const T

returns a raw pointer to the element at the given index, assuming the provided index is within the column bounds

safety

the behavior is undefined if any of the following conditions are violated:

row < self.nrows()

pub fn split_at_row( self, row: IdxInc<Rows>, ) -> (ColRef<'a, T, usize, RStride>, ColRef<'a, T, usize, RStride>)

splits the column horizontally at the given row into two parts and returns an array of each submatrix, in the following order:

top

bottom

panics

the function panics if the following condition is violated:

row <= self.nrows()

pub fn transpose(self) -> RowRef<'a, T, Rows, RStride>

returns a view over the transpose of self

pub fn conjugate(self) -> ColRef<'a, T::Conj, Rows, RStride>
where T: Conjugate,

returns a view over the conjugate of self

pub fn canonical(self) -> ColRef<'a, T::Canonical, Rows, RStride>
where T: Conjugate,

returns an unconjugated view over self

pub fn adjoint(self) -> RowRef<'a, T::Conj, Rows, RStride>
where T: Conjugate,

returns a view over the conjugate transpose of self

pub fn get<RowRange>( self, row: RowRange, ) -> <ColRef<'a, T, Rows, RStride> as ColIndex<RowRange>>::Target
where ColRef<'a, T, Rows, RStride>: ColIndex<RowRange>,

returns a reference to the element at the given index, or a subcolumn if row is a range

panics

the function panics if any of the following conditions are violated:

row must be contained in [0, self.nrows())

pub unsafe fn get_unchecked<RowRange>( self, row: RowRange, ) -> <ColRef<'a, T, Rows, RStride> as ColIndex<RowRange>>::Target
where ColRef<'a, T, Rows, RStride>: ColIndex<RowRange>,

returns a reference to the element at the given index, or a subcolumn if row is a range, without bound checks

safety

the behavior is undefined if any of the following conditions are violated:

row must be contained in [0, self.nrows())

pub fn reverse_rows(self) -> ColRef<'a, T, Rows, RStride::Rev>

returns a view over the self, with the rows in reversed order

pub fn subrows<V: Shape>( self, row_start: IdxInc<Rows>, nrows: V, ) -> ColRef<'a, T, V, RStride>

returns a view over the column starting at row row_start, and with number of rows nrows

panics

the function panics if any of the following conditions are violated:

row_start <= self.nrows()

nrows <= self.nrows() - row_start

pub fn as_row_shape<V: Shape>(self, nrows: V) -> ColRef<'a, T, V, RStride>

returns the input column with the given row shape after checking that it matches the current row shape

pub fn as_dyn_rows(self) -> ColRef<'a, T, usize, RStride>

returns the input column with dynamic row shape

pub fn as_dyn_stride(self) -> ColRef<'a, T, Rows, isize>

returns the input column with dynamic stride

pub fn iter( self, ) -> impl 'a + ExactSizeIterator + DoubleEndedIterator<Item = &'a T>
where Rows: 'a,

returns an iterator over the elements of the column

pub fn par_iter(self) -> impl 'a + IndexedParallelIterator<Item = &'a T>
where T: Sync, Rows: 'a,

returns a parallel iterator over the elements of the column

pub fn par_partition( self, count: usize, ) -> impl 'a + IndexedParallelIterator<Item = ColRef<'a, T, usize, RStride>>
where T: Sync, Rows: 'a,

returns a parallel iterator that provides exactly count successive chunks of the elements of this column

only available with the rayon feature

pub fn try_as_col_major(self) -> Option<ColRef<'a, T, Rows, ContiguousFwd>>

returns a view over the column with a static row stride equal to +1, or None otherwise

pub fn as_mat(self) -> MatRef<'a, T, Rows, usize, RStride, isize>

returns a matrix view over self

pub fn as_diagonal(self) -> DiagRef<'a, T, Rows, RStride>

interprets the column as a diagonal matrix

impl<'a, T, Rows: Shape> ColRef<'a, T, Rows, ContiguousFwd>

pub fn as_slice(self) -> &'a [T]

returns a reference over the elements as a slice

impl<'a, 'ROWS, T> ColRef<'a, T, Dim<'ROWS>, ContiguousFwd>

pub fn as_array(self) -> &'a Array<'ROWS, T>

returns a reference over the elements as a lifetime-bound slice

impl<'a, T, Rows: Shape, RStride: Stride> ColRef<'a, T, Rows, RStride>
where T: RealField,

pub fn max(&self) -> Option<T>

Returns the maximum element in the column, or None if the column is empty

pub fn min(&self) -> Option<T>

Returns the minimum element in the column, or None if the column is empty