faer::diag::generic

Struct Diag

#[repr(transparent)]
pub struct Diag<Inner>(pub Inner);

generic Diag wrapper

Tuple Fields
0: Inner
Implementations

impl<'a, T> Diag<Mut<'a, T>>

pub fn from_mut(value: &'a mut T) -> Self

creates a diagonal matrix view over the given element

pub fn from_slice_mut(slice: &'a mut [T]) -> Self

creates a DiagMut from slice views over the diagonal data, the result has the same dimension as the length of the input slice

impl<'a, T, Dim: Shape, Stride: Stride> Diag<Mut<'a, T, Dim, Stride>>

pub const unsafe fn from_raw_parts_mut( ptr: *mut T, dim: Dim, stride: Stride, ) -> Self

creates a DiagMut from pointers to the diagonal data, dimension, and stride

safety

this function has the same safety requirements as [MatMut::from_raw_parts_mut(ptr, dim, 1, stride, 0)]

pub fn column_vector(self) -> ColRef<'a, T, Dim, Stride>

returns the diagonal as a column vector view

pub fn column_vector_mut(self) -> ColMut<'a, T, Dim, Stride>

returns the diagonal as a mutable column vector view

pub fn as_ref(&self) -> DiagRef<'_, T, Dim, Stride>

returns a view over self

pub fn as_mut(&mut self) -> DiagMut<'_, T, Dim, Stride>

returns a view over self

pub fn fill(&mut self, value: T)
where T: Clone,

fills all the elements of self with value

pub fn as_shape<D: Shape>(self, len: D) -> DiagRef<'a, T, D, Stride>

see DiagRef::as_shape

pub fn as_dyn(self) -> DiagRef<'a, T, usize, Stride>

see DiagRef::as_dyn

pub fn as_dyn_stride(self) -> DiagRef<'a, T, Dim>

see DiagRef::as_dyn_stride

pub fn conjugate(self) -> DiagRef<'a, T::Conj, Dim, Stride>
where T: Conjugate,

see DiagRef::conjugate

pub fn canonical(self) -> DiagRef<'a, T::Canonical, Dim, Stride>
where T: Conjugate,

see DiagRef::canonical

pub fn as_shape_mut<D: Shape>(self, len: D) -> DiagMut<'a, T, D, Stride>

see DiagRef::as_shape

pub fn as_dyn_mut(self) -> DiagMut<'a, T, usize, Stride>

see DiagRef::as_dyn

pub fn as_dyn_stride_mut(self) -> DiagMut<'a, T, Dim>

see DiagRef::as_dyn_stride

pub fn conjugate_mut(self) -> DiagMut<'a, T::Conj, Dim, Stride>
where T: Conjugate,

see DiagRef::conjugate

pub fn canonical_mut(self) -> DiagMut<'a, T::Canonical, Dim, Stride>
where T: Conjugate,

see DiagRef::canonical

pub fn dim(&self) -> Dim

returns the dimension of self

pub fn copy_from<RhsT: Conjugate<Canonical = T>>( &mut self, rhs: impl AsDiagRef<T = RhsT, Dim = Dim>, )
where T: ComplexField,

copies other into self

impl<T, Dim: Shape> Diag<Own<T, Dim>>

pub fn column_vector(&self) -> ColRef<'_, T, Dim>

returns the diagonal as a column vector

pub fn column_vector_mut(&mut self) -> ColMut<'_, T, Dim>

returns the diagonal as a column vector

pub fn into_column_vector(self) -> Col<T, Dim>

returns the diagonal as a column vector

pub fn as_ref(&self) -> DiagRef<'_, T, Dim>

returns a view over self

pub fn as_mut(&mut self) -> DiagMut<'_, T, Dim>

returns a view over self

pub fn as_shape<D: Shape>(&self, len: D) -> DiagRef<'_, T, D>

see DiagRef::as_shape

pub fn as_dyn(&self) -> DiagRef<'_, T>

see DiagRef::as_dyn

pub fn as_dyn_mut(&mut self) -> DiagMut<'_, T>

see DiagMut::as_dyn_mut

pub fn conjugate(&self) -> DiagRef<'_, T::Conj, Dim>
where T: Conjugate,

see DiagRef::conjugate

pub fn canonical(&self) -> DiagRef<'_, T::Canonical, Dim>
where T: Conjugate,

see DiagRef::canonical

pub fn as_shape_mut<D: Shape>(&mut self, len: D) -> DiagMut<'_, T, D>

see DiagMut::as_shape_mut

pub fn conjugate_mut(&mut self) -> DiagMut<'_, T::Conj, Dim>
where T: Conjugate,

see DiagMut::conjugate_mut

pub fn canonical_mut(&mut self) -> DiagMut<'_, T::Canonical, Dim>
where T: Conjugate,

see DiagMut::canonical_mut

pub fn dim(&self) -> Dim

returns the dimension of self

pub fn zeros(dim: Dim) -> Self
where T: ComplexField,

returns a new diagonal with dimension dim, filled with zeros

pub fn ones(dim: Dim) -> Self
where T: ComplexField,

returns a new diagonal with dimension dim, filled with ones

pub fn full(dim: Dim, value: T) -> Self
where T: Clone,

returns a new diagonal with dimension dim, filled with value

pub fn copy_from<RhsT: Conjugate<Canonical = T>>( &mut self, rhs: impl AsDiagRef<T = RhsT, Dim = Dim>, )
where T: ComplexField,

copies other into self

impl<'a, T> Diag<Ref<'a, T>>

pub fn from_ref(value: &'a T) -> Self

creates a diagonal matrix view over the given element

pub fn from_slice(slice: &'a [T]) -> Self

creates a DiagRef from slice views over the diagonal data, the result has the same dimension as the length of the input slice

impl<'a, T, Dim: Shape, Stride: Stride> Diag<Ref<'a, T, Dim, Stride>>

pub const unsafe fn from_raw_parts( ptr: *const T, dim: Dim, stride: Stride, ) -> Self

creates a DiagRef from pointers to the diagonal data, dimension, and stride

safety

this function has the same safety requirements as [MatRef::from_raw_parts(ptr, dim, 1, stride, 0)]

pub fn column_vector(self) -> ColRef<'a, T, Dim, Stride>

returns the diagonal as a column vector view.

pub fn as_ref(&self) -> DiagRef<'_, T, Dim, Stride>

returns a view over self

pub fn as_shape<D: Shape>(self, len: D) -> DiagRef<'a, T, D, Stride>

returns the input matrix with the given shape after checking that it matches the current shape

pub fn as_dyn(self) -> DiagRef<'a, T, usize, Stride>

returns the input matrix with dynamic shape

pub fn as_dyn_stride(self) -> DiagRef<'a, T, Dim>

returns the input matrix with dynamic stride

pub fn conjugate(self) -> DiagRef<'a, T::Conj, Dim, Stride>
where T: Conjugate,

returns a view over the conjugate of self

pub fn canonical(self) -> DiagRef<'a, T::Canonical, Dim, Stride>
where T: Conjugate,

returns an unconjugated view over self

pub fn dim(&self) -> Dim

returns the dimension of self

impl<T, Dim: Shape, Stride: Stride, Inner: for<'short> Reborrow<'short, Target = Ref<'short, T, Dim, Stride>>> Diag<Inner>

pub fn is_all_finite(&self) -> bool
where T: Conjugate,

returns true if all of the elements of self are finite. otherwise returns false.

pub fn has_nan(&self) -> bool
where T: Conjugate,

returns true if any of the elements of self is NaN. otherwise returns false.

impl<Inner> Diag<Inner>

pub fn from_inner_ref(inner: &Inner) -> &Self

wrap by reference

pub fn from_inner_mut(inner: &mut Inner) -> &mut Self

wrap by mutable reference