faer::diag

Type Alias Diag

pub type Diag<T, Dim = usize> = Diag<Own<T, Dim>>;

diagonal matrix

Aliased Type

#[repr(transparent)]
pub struct Diag<T, Dim = usize>(pub Own<T, Dim>);

Tuple Fields
0: Own<T, Dim>
Implementations

impl<T, Dim: Shape> Diag<T, Dim>

pub fn column_vector(&self) -> ColRef<'_, T, Dim>

returns the diagonal as a column vector

pub fn column_vector_mut(&mut self) -> ColMut<'_, T, Dim>

returns the diagonal as a column vector

pub fn into_column_vector(self) -> Col<T, Dim>

returns the diagonal as a column vector

pub fn as_ref(&self) -> DiagRef<'_, T, Dim>

returns a view over self

pub fn as_mut(&mut self) -> DiagMut<'_, T, Dim>

returns a view over self

pub fn as_shape<D: Shape>(&self, len: D) -> DiagRef<'_, T, D>

see DiagRef::as_shape

pub fn as_dyn(&self) -> DiagRef<'_, T>

see DiagRef::as_dyn

pub fn as_dyn_mut(&mut self) -> DiagMut<'_, T>

see DiagMut::as_dyn_mut

pub fn conjugate(&self) -> DiagRef<'_, T::Conj, Dim>
where T: Conjugate,

see DiagRef::conjugate

pub fn canonical(&self) -> DiagRef<'_, T::Canonical, Dim>
where T: Conjugate,

see DiagRef::canonical

pub fn as_shape_mut<D: Shape>(&mut self, len: D) -> DiagMut<'_, T, D>

see DiagMut::as_shape_mut

pub fn conjugate_mut(&mut self) -> DiagMut<'_, T::Conj, Dim>
where T: Conjugate,

see DiagMut::conjugate_mut

pub fn canonical_mut(&mut self) -> DiagMut<'_, T::Canonical, Dim>
where T: Conjugate,

see DiagMut::canonical_mut

pub fn dim(&self) -> Dim

returns the dimension of self

pub fn zeros(dim: Dim) -> Self
where T: ComplexField,

returns a new diagonal with dimension dim, filled with zeros

pub fn ones(dim: Dim) -> Self
where T: ComplexField,

returns a new diagonal with dimension dim, filled with ones

pub fn full(dim: Dim, value: T) -> Self
where T: Clone,

returns a new diagonal with dimension dim, filled with value

pub fn copy_from<RhsT: Conjugate<Canonical = T>>( &mut self, rhs: impl AsDiagRef<T = RhsT, Dim = Dim>, )
where T: ComplexField,

copies other into self