faer::diag

Type Alias DiagRef

pub type DiagRef<'a, T, Dim = usize, Stride = isize> = Diag<Ref<'a, T, Dim, Stride>>;

diagonal matrix view

Aliased Type

#[repr(transparent)]
pub struct DiagRef<'a, T, Dim = usize, Stride = isize>(pub Ref<'a, T, Dim, Stride>);

Tuple Fields
0: Ref<'a, T, Dim, Stride>
Implementations

impl<'a, T> DiagRef<'a, T>

pub fn from_ref(value: &'a T) -> Self

creates a diagonal matrix view over the given element

pub fn from_slice(slice: &'a [T]) -> Self

creates a DiagRef from slice views over the diagonal data, the result has the same dimension as the length of the input slice

impl<'a, T, Dim: Shape, Stride: Stride> DiagRef<'a, T, Dim, Stride>

pub const unsafe fn from_raw_parts( ptr: *const T, dim: Dim, stride: Stride, ) -> Self

creates a DiagRef from pointers to the diagonal data, dimension, and stride

safety

this function has the same safety requirements as [MatRef::from_raw_parts(ptr, dim, 1, stride, 0)]

pub fn column_vector(self) -> ColRef<'a, T, Dim, Stride>

returns the diagonal as a column vector view.

pub fn as_ref(&self) -> DiagRef<'_, T, Dim, Stride>

returns a view over self

pub fn as_shape<D: Shape>(self, len: D) -> DiagRef<'a, T, D, Stride>

returns the input matrix with the given shape after checking that it matches the current shape

pub fn as_dyn(self) -> DiagRef<'a, T, usize, Stride>

returns the input matrix with dynamic shape

pub fn as_dyn_stride(self) -> DiagRef<'a, T, Dim>

returns the input matrix with dynamic stride

pub fn conjugate(self) -> DiagRef<'a, T::Conj, Dim, Stride>
where T: Conjugate,

returns a view over the conjugate of self

pub fn canonical(self) -> DiagRef<'a, T::Canonical, Dim, Stride>
where T: Conjugate,

returns an unconjugated view over self

pub fn dim(&self) -> Dim

returns the dimension of self