Module faer::iter

Iterators and related utilities.

Structs

ColChunks

Chunked matrix iterator over the columns.

ColChunksMut

Chunked matrix iterator over the columns.

ColElemChunks

Chunked column iterator over the elements.

ColElemChunksMut

Chunked column iterator over the elements.

ColElemPartition

Fixed-size chunked column iterator over the elements.

ColElemPartitionMut

Fixed-size chunked column iterator over the elements.

ColIter

Iterator over the columns of a matrix.

ColIterMut

Iterator over the columns of a matrix.

ColPartition

Fixed-size chunked matrix iterator over the columns.

ColPartitionMut

Fixed-size chunked matrix iterator over the columns.

ElemIter

Iterator over the elements of a row or column.

ElemIterMut

Iterator over the elements of a row or column.

RowChunks

Chunked matrix iterator over the rows.

RowChunksMut

Chunked matrix iterator over the rows.

RowElemChunks

Chunked row iterator over the elements.

RowElemChunksMut

Chunked row iterator over the elements.

RowElemPartition

Fixed-size chunked row iterator over the elements.

RowElemPartitionMut

Fixed-size chunked row iterator over the elements.

RowIter

Iterator over the rows of a matrix.

RowIterMut

Iterator over the rows of a matrix.

RowPartition

Fixed-size chunked matrix iterator over the rows.

RowPartitionMut

Fixed-size chunked matrix iterator over the rows.