Module faer::linalg

Linear algebra module.

Contains low level routines and the implementation of their corresponding high level wrappers.

Memory allocation

Since most faer crates aim to expose a low level api for optimal performance, most algorithms try to defer memory allocation to the user.

However, since a lot of algorithms need some form of temporary space for intermediate computations, they may ask for a slice of memory for that purpose, by taking a stack: PodStack parameter. A PodStack is a thin wrapper over a slice of memory bytes. This memory may come from any valid source (heap allocation, fixed-size array on the stack, etc.). The functions taking a PodStack parameter have a corresponding function with a similar name ending in _req that returns the memory requirements of the algorithm. For example: householder::apply_block_householder_on_the_left_in_place_with_conj and householder::apply_block_householder_on_the_left_in_place_req.

The memory stack may be reused in user-code to avoid repeated allocations, and it is also possible to compute the sum (dyn_stack::StackReq::all_of) or union (dyn_stack::StackReq::any_of) of multiple requirements, in order to optimally combine them into a single allocation.

After computing a dyn_stack::StackReq, one can query its size and alignment to allocate the required memory. The simplest way to do so is through dyn_stack::GlobalMemBuffer::new.

Entity trait

Matrices are built on top of the Entity trait, which describes the preferred memory storage layout for a given type E. An entity can be decomposed into a group of units: for a natively supported type (f32, f64, c32, c64), the unit is simply the type itself, and a group contains a single element. On the other hand, for a type with a more specific preferred layout, like an extended precision floating point type, or a dual number type, the unit would be one of the natively supported types, and the group would be a structure holding the components that build up the full value.

To take a more specific example: num_complex::Complex<f64> has a storage memory layout that differs from that of c64 (see faer::complex_native for more details). Its real and complex components are stored separately, so its unit type is f64, while its group type is Complex. In practice, this means that for a Mat<f64>, methods such as Mat::col_as_slice will return a &[f64]. Meanwhile, for a Mat<Complex<f64>>, Mat::col_as_slice will return Complex<&[f64]>, which holds two slices, each pointing respectively to a view over the real and the imaginary components.

While the design of the entity trait is unconventional, it helps us achieve much higher performance when targeting non native types, due to the design matching the typical preferred CPU layout for SIMD operations. And for native types, since Group<T> is just T, the entity layer is a no-op, and the matrix layout is compatible with the classic contiguous layout that’s commonly used by other libraries.

Re-exports

pub use faer_entity as entity;

Modules

cholesky

Low level implementation of the various Cholesky-like decompositions.

evd

Low level implementation of the eigenvalue decomposition of a square diagonalizable matrix.

householder

Block Householder transformations.

lu

Low level implementation of the LU decompositions.

matmul

Matrix multiplication.

qr

This crate provides utilities for computing and manipulating the QR factorization with and without pivoting. The QR factorization decomposes a matrix into a product of a unitary matrix $Q$ (represented using block Householder sequences), and an upper trapezoidal matrix $R$, such that their product is equal to the original matrix (or a column permutation of it in the case where column pivoting is used).

solvers

High level linear system solvers.

svd

Low level implementation of the SVD of a matrix.

triangular_inverse

Triangular matrix inversion.

triangular_solve

Triangular solve module.

zip

Implementation of zipped! structures.

Functions

kron

Kronecker product of two matrices.

temp_mat_constant

Creates a temporary matrix of constant values, from the given memory stack.

temp_mat_req

Returns the stack requirements for creating a temporary matrix with the given dimensions.

temp_mat_uninit

Creates a temporary matrix of untouched values, from the given memory stack.

temp_mat_zeroed

Creates a temporary matrix of zero values, from the given memory stack.