Module faer::linalg::cholesky

Low level implementation of the various Cholesky-like decompositions.

Modules

bunch_kaufman

The Bunch Kaufman decomposition of a Hermitian matrix $A$ is such that: $$P A P^\top = LBL^H,$$ where $P$ is a permutation matrix, $B$ is a block diagonal matrix, with $1\times 1$ or $2 \times 2 $ diagonal blocks, and $L$ is a unit lower triangular matrix.

ldlt_diagonal

The Cholesky decomposition with diagonal $D$ of a Hermitian matrix $A$ is such that: $$A = LDL^H,$$ where $D$ is a diagonal matrix, and $L$ is a unit lower triangular matrix.

llt

The Cholesky decomposition of a Hermitian positive definite matrix $A$ is such that: $$A = LL^H,$$ where $L$ is a lower triangular matrix.

Functions

compute_cholesky_permutation

Computes a permutation that reduces the chance of numerical errors during the $LDL^H$ factorization with diagonal $D$, then stores the result in perm_indices and perm_inv_indices.