faer::linalg::cholesky::lblt::factor

Function cholesky_in_place

pub fn cholesky_in_place<'out, I: Index, T: ComplexField>(
    A: MatMut<'_, T>,
    subdiag: DiagMut<'_, T>,
    perm: &'out mut [I],
    perm_inv: &'out mut [I],
    par: Par,
    stack: &mut MemStack,
    params: Spec<LbltParams, T>,
) -> (LbltInfo, PermRef<'out, I>)

computes the $LBL^\top$ factorization of $A$ and stores the factorization in matrix and subdiag

the diagonal of the block diagonal matrix is stored on the diagonal of matrix, while the subdiagonal elements of the blocks are stored in subdiag

panics

panics if the input matrix is not square

this can also panic if the provided memory in stack is insufficient (see cholesky_in_place_scratch).