faer::linalg::cholesky::lblt::solve

Function solve_in_place_with_conj

pub fn solve_in_place_with_conj<I: Index, T: ComplexField>(
    L: MatRef<'_, T>,
    diagonal: DiagRef<'_, T>,
    subdiagonal: DiagRef<'_, T>,
    conj_A: Conj,
    perm: PermRef<'_, I>,
    rhs: MatMut<'_, T>,
    par: Par,
    stack: &mut MemStack,
)

given the bunch-kaufman factors of a matrix $a$ and a matrix $b$ stored in rhs, this function computes the solution of the linear system $A x = b$, implicitly conjugating $A$ if needed

the solution of the linear system is stored in rhs

panics

panics if lb_factors is not a square matrix

panics if subdiag is not a column vector with the same number of rows as the dimension of lb_factors

panics if rhs doesn’t have the same number of rows as the dimension of lb_factors

panics if the provided memory in stack is insufficient (see solve_in_place_scratch)