faer::linalg::householder

Function apply_block_householder_transpose_on_the_left_in_place_scratch

pub fn apply_block_householder_transpose_on_the_left_in_place_scratch<T: ComplexField>(
    householder_basis_nrows: usize,
    blocksize: usize,
    rhs_ncols: usize,
) -> StackReq

computes the size and alignment of required workspace for applying the transpose of a block householder transformation to a right-hand-side matrix in place