faer::linalg::qr::no_pivoting

Module factor

Structs

QrInfo

information about the resulting $QR$ factorization.

QrParams

$QR$ factorization tuning parameters.

Functions

qr_in_place

qr_in_place_scratch

computes the size and alignment of required workspace for performing a qr decomposition with no pivoting

recommended_blocksize

the recommended block size to use for a $QR$ decomposition of a matrix with the given shape.