Module faer::linalg::solvers

High level linear system solvers.

Re-exports

pub use crate::linalg::cholesky::llt::CholeskyError;

pub use crate::sparse::linalg::solvers::SpSolver;

pub use crate::sparse::linalg::solvers::SpSolverCore;

pub use crate::sparse::linalg::solvers::SpSolverLstsq;

pub use crate::sparse::linalg::solvers::SpSolverLstsqCore;

Structs

Cholesky

Cholesky decomposition.

ColPivQr

QR decomposition with column pivoting.

Eigendecomposition

Complex eigendecomposition.

FullPivLu

LU decomposition with full pivoting.

Lblt

Bunch-Kaufman decomposition.

PartialPivLu

LU decomposition with partial pivoting.

Qr

QR decomposition.

SelfAdjointEigendecomposition

Self-adjoint eigendecomposition.

Svd

Singular value decomposition.

ThinSvd

Thin singular value decomposition.

Traits

Solver

Solver that can compute solution of a linear system.

SolverCore

Object-safe base for Solver

SolverLstsq

Dense solver that can compute the least squares solution of an overdetermined linear system.

SolverLstsqCore

Object-safe base for SolverLstsq