Struct faer::linalg::solvers::FullPivLu

pub struct FullPivLu<E: Entity> { /* private fields */ }

LU decomposition with full pivoting.

Implementations

impl<E: ComplexField> FullPivLu<E>

pub fn new<ViewE: Conjugate<Canonical = E>>(matrix: MatRef<'_, ViewE>) -> Self

Returns the LU decomposition of the input matrix with row and column pivoting.

The factorization is such that $PAQ^\top = LU$, where $L$ is lower triangular, $U$ is unit upper triangular, and $P$ is the permutation arising from row pivoting and $Q$ is the permutation due to column pivoting.

pub fn row_permutation(&self) -> PermRef<'_, usize>

Returns the row permutation due to pivoting.

pub fn col_permutation(&self) -> PermRef<'_, usize>

Returns the column permutation due to pivoting.

pub fn transposition_count(&self) -> usize

Returns the number of transpositions that constitute the two permutations.

pub fn compute_l(&self) -> Mat<E>

Returns the factor $L$ of the LU decomposition.

pub fn compute_u(&self) -> Mat<E>

Returns the factor $U$ of the LU decomposition.