faer::linalg::solvers

Trait DenseSolveCore

pub trait DenseSolveCore<T: ComplexField>: SolveCore<T> {
    // Required methods
    fn reconstruct(&self) -> Mat<T>;
    fn inverse(&self) -> Mat<T>;
}

dense linear system solver

Required Methods

fn reconstruct(&self) -> Mat<T>

returns an approximation of the matrix that was used to create the decomposition

fn inverse(&self) -> Mat<T>

returns an approximation of the inverse of the matrix that was used to create the decomposition