faer::linalg::solvers

Trait SolveCore

pub trait SolveCore<T: ComplexField>: ShapeCore {
    // Required methods
    fn solve_in_place_with_conj(&self, conj: Conj, rhs: MatMut<'_, T>);
    fn solve_transpose_in_place_with_conj(&self, conj: Conj, rhs: MatMut<'_, T>);
}

linear system solver implementation

Required Methods

fn solve_in_place_with_conj(&self, conj: Conj, rhs: MatMut<'_, T>)

solves the equation self × x = rhs, implicitly conjugating self if needed, and stores the result in rhs

fn solve_transpose_in_place_with_conj(&self, conj: Conj, rhs: MatMut<'_, T>)

solves the equation self.transpose() × x = rhs, implicitly conjugating self if needed, and stores the result in rhs