faer::linalg::solvers

Trait SolveLstsqCore

pub trait SolveLstsqCore<T: ComplexField>: ShapeCore {
    // Required method
    fn solve_lstsq_in_place_with_conj(&self, conj: Conj, rhs: MatMut<'_, T>);
}

least squares linear system solver implementation

Required Methods

fn solve_lstsq_in_place_with_conj(&self, conj: Conj, rhs: MatMut<'_, T>)

solves the equation self × x = rhs in the sense of least squares, implicitly conjugating self if needed, and stores the result in the top rows of rhs