faer::linalg::zip

Struct ZipEq

pub struct ZipEq<Rows, Cols, Head, Tail>(pub Head, pub Tail, _);

zipped matrices

Tuple Fields
0: Head1: Tail
Implementations

impl<Rows: Copy + Eq + Debug, Cols: Copy + Eq + Debug, Head: MatIndex<Rows = Rows, Cols = Cols>, Tail: MatIndex<Rows = Rows, Cols = Cols>> ZipEq<Rows, Cols, Head, Tail>

single matrix view

pub fn new(head: Head, tail: Tail) -> Self

creates a zip matrix, after asserting that the dimensions match

pub fn new_unchecked(head: Head, tail: Tail) -> Self

creates a zip matrix, assuming that the dimensions match

impl<Rows: Shape, Cols: Shape, L: MatIndex<LayoutTransform = MatLayoutTransform, Rows = Rows, Cols = Cols, Index = (Idx<Rows>, Idx<Cols>)>, R: MatIndex<LayoutTransform = MatLayoutTransform, Rows = Rows, Cols = Cols, Index = (Idx<Rows>, Idx<Cols>)>> ZipEq<Rows, Cols, L, R>
where L::Dyn: MatIndex<Rows = usize, Cols = usize, Index = (usize, usize)>, R::Dyn: MatIndex<Rows = usize, Cols = usize, Index = (usize, usize)>,

pub fn for_each(self, f: impl FnMut(<Self as MatIndex>::Item))

applies f to each element of self

pub fn for_each_with_index( self, f: impl FnMut(Idx<Rows>, Idx<Cols>, <Self as MatIndex>::Item), )

applies f to each element of self, while passing the indices of the position of the current element

pub fn for_each_triangular_lower_with_index( self, diag: Diag, f: impl FnMut(Idx<Rows>, Idx<Cols>, <Self as MatIndex>::Item), )

applies f to each element of the lower triangular half of self, while passing the indices of the position of the current element

diag specifies whether the diagonal should be included or excluded

pub fn for_each_triangular_upper_with_index( self, diag: Diag, f: impl FnMut(Idx<Rows>, Idx<Cols>, <Self as MatIndex>::Item), )

applies f to each element of the upper triangular half of self, while passing the indices of the position of the current element

diag specifies whether the diagonal should be included or excluded

pub fn for_each_triangular_lower( self, diag: Diag, f: impl FnMut(<Self as MatIndex>::Item), )

applies f to each element of the lower triangular half of self

diag specifies whether the diagonal should be included or excluded

pub fn for_each_triangular_upper( self, diag: Diag, f: impl FnMut(<Self as MatIndex>::Item), )

applies f to each element of the upper triangular half of self

diag specifies whether the diagonal should be included or excluded

pub fn map<T>( self, f: impl FnMut(<Self as MatIndex>::Item) -> T, ) -> Mat<T, Rows, Cols>

applies f to each element of self and collect its result into a new matrix

pub fn map_with_index<T>( self, f: impl FnMut(Idx<Rows>, Idx<Cols>, <Self as MatIndex>::Item) -> T, ) -> Mat<T, Rows, Cols>

applies f to each element of self and collect its result into a new matrix

impl<Rows: Shape, L: MatIndex<LayoutTransform = VecLayoutTransform, Rows = Rows, Cols = (), Index = Idx<Rows>>, R: MatIndex<LayoutTransform = VecLayoutTransform, Rows = Rows, Cols = (), Index = Idx<Rows>>> ZipEq<Rows, (), L, R>
where L::Dyn: MatIndex<Rows = usize, Cols = (), Index = usize>, R::Dyn: MatIndex<Rows = usize, Cols = (), Index = usize>,

pub fn for_each(self, f: impl FnMut(<Self as MatIndex>::Item))

applies f to each element of self

pub fn for_each_with_index( self, f: impl FnMut(Idx<Rows>, <Self as MatIndex>::Item), )

applies f to each element of self, while passing the indices of the position of the current element

pub fn map<T>( self, f: impl FnMut(<Self as MatIndex>::Item) -> T, ) -> Col<T, Rows>

applies f to each element of self and collect its result into a new matrix

pub fn map_with_index<T>( self, f: impl FnMut(Idx<Rows>, <Self as MatIndex>::Item) -> T, ) -> Col<T, Rows>

applies f to each element of self and collect its result into a new matrix

impl<Cols: Shape, L: MatIndex<LayoutTransform = VecLayoutTransform, Rows = (), Cols = Cols, Index = Idx<Cols>>, R: MatIndex<LayoutTransform = VecLayoutTransform, Rows = (), Cols = Cols, Index = Idx<Cols>>> ZipEq<(), Cols, L, R>
where L::Dyn: MatIndex<Rows = (), Cols = usize, Index = usize>, R::Dyn: MatIndex<Rows = (), Cols = usize, Index = usize>,

pub fn for_each(self, f: impl FnMut(<Self as MatIndex>::Item))

applies f to each element of self

pub fn for_each_with_index( self, f: impl FnMut(Idx<Cols>, <Self as MatIndex>::Item), )

applies f to each element of self, while passing the indices of the position of the current element

pub fn map<T>( self, f: impl FnMut(<Self as MatIndex>::Item) -> T, ) -> Row<T, Cols>

applies f to each element of self and collect its result into a new matrix

pub fn map_with_index<T>( self, f: impl FnMut(Idx<Cols>, <Self as MatIndex>::Item) -> T, ) -> Row<T, Cols>

applies f to each element of self and collect its result into a new matrix