faer::linalg::zip

Trait MatIndex

pub trait MatIndex {
    type Rows: Copy + Eq + Debug;
    type Cols: Copy + Eq + Debug;
    type Index: Copy;
    type LayoutTransform: Copy;
    type Item;
    type Dyn: MatIndex<Dyn = Self::Dyn, LayoutTransform = Self::LayoutTransform, Item = Self::Item, Slice = Self::Slice>;
    type Slice: for<'a> SliceFamily<'a, Self::Item>;

    // Required methods
    fn nrows(this: &Self) -> Self::Rows;
    fn ncols(this: &Self) -> Self::Cols;
    unsafe fn get_slice_unchecked<'a>(
        this: &'a mut Self,
        idx: Self::Index,
        n_elems: usize,
    ) -> <Self::Slice as SliceFamily<'a, Self::Item>>::Slice;
    unsafe fn from_dyn_idx(idx: <Self::Dyn as MatIndex>::Index) -> Self::Index;
    unsafe fn get_unchecked(this: &mut Self, index: Self::Index) -> Self::Item;
    unsafe fn next_unchecked<'a>(
        slice: &mut <Self::Slice as SliceFamily<'a, Self::Item>>::Slice,
    ) -> Self::Item;
    fn is_contiguous(this: &Self) -> bool;
    fn preferred_layout(this: &Self) -> Self::LayoutTransform;
    fn with_layout(this: Self, layout: Self::LayoutTransform) -> Self::Dyn;
}

type with a given matrix shape

Required Associated Types

type Rows: Copy + Eq + Debug

type of rows

type Cols: Copy + Eq + Debug

type of columns

type Index: Copy

indexing type

type LayoutTransform: Copy

layout transformation type

type Item

item produced by the zip views

type Dyn: MatIndex<Dyn = Self::Dyn, LayoutTransform = Self::LayoutTransform, Item = Self::Item, Slice = Self::Slice>

matrix type with type erased dimensions

type Slice: for<'a> SliceFamily<'a, Self::Item>

Required Methods

fn nrows(this: &Self) -> Self::Rows

returns the number of rows

fn ncols(this: &Self) -> Self::Cols

returns the number of columns

unsafe fn get_slice_unchecked<'a>( this: &'a mut Self, idx: Self::Index, n_elems: usize, ) -> <Self::Slice as SliceFamily<'a, Self::Item>>::Slice

returns slice at index of length n_elems

unsafe fn from_dyn_idx(idx: <Self::Dyn as MatIndex>::Index) -> Self::Index

converts a type erased index back to its original representation

unsafe fn get_unchecked(this: &mut Self, index: Self::Index) -> Self::Item

get the item at the given index, skipping bound checks

unsafe fn next_unchecked<'a>( slice: &mut <Self::Slice as SliceFamily<'a, Self::Item>>::Slice, ) -> Self::Item

get the item at the given slice position, skipping bound checks

fn is_contiguous(this: &Self) -> bool

checks if the zip matrices are contiguous

fn preferred_layout(this: &Self) -> Self::LayoutTransform

computes the preferred iteration layout of the matrices

fn with_layout(this: Self, layout: Self::LayoutTransform) -> Self::Dyn

applies the layout transformation to the matrices

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.