faer::matrix_free::eigen

Function partial_eigen

pub fn partial_eigen<T: ComplexField>(
    eigvecs: MatMut<'_, Complex<T::Real>>,
    eigvals: &mut [Complex<T::Real>],
    A: &dyn LinOp<T>,
    v0: ColRef<'_, T>,
    tolerance: T::Real,
    par: Par,
    stack: &mut MemStack,
    params: PartialEigenParams,
) -> PartialEigenInfo

computes an estimate of the eigenvalues (and corresponding eigenvectors) of $A$ with the largest magnitude until the provided outputs are full or the maximum number of algorithm restarts is reached.