Module faer::modules::core

👎Deprecated

Emulation layer for faer_core

Re-exports

pub use crate::col;

pub use crate::col;

pub use crate::complex_native;

pub use crate::utils::constrained;

pub use crate::utils as group_helpers;

pub use crate::linalg::householder;

pub use crate::linalg::triangular_inverse as inverse;

pub use crate::linalg::triangular_inverse as solve;

pub use crate::mat;

pub use crate::mat;

pub use crate::linalg::matmul as mul;

pub use crate::perm as permutation;

pub use crate::row;

pub use crate::row;

pub use crate::sparse;

pub use crate::linalg::zip;

pub use crate::Conj;

pub use crate::sparse::FaerError;

pub use crate::Parallelism;

pub use crate::Side;

pub use crate::col::AsColRef;

pub use crate::col::AsColMut;

pub use crate::row::AsRowRef;

pub use crate::row::AsRowMut;

pub use crate::mat::AsMatRef;

pub use crate::mat::AsMatMut;

pub use crate::mat::As2D;

pub use crate::mat::As2DMut;

pub use crate::linalg::kron;

pub use crate::disable_global_parallelism;

pub use crate::set_global_parallelism;

pub use crate::get_global_parallelism;

pub use crate::linalg::temp_mat_req;

pub use crate::linalg::temp_mat_uninit;

pub use crate::linalg::temp_mat_constant;

pub use crate::linalg::temp_mat_zeroed;

pub use crate::Col;

pub use crate::ColMut;

pub use crate::ColRef;

pub use crate::Row;

pub use crate::RowMut;

pub use crate::RowRef;

pub use crate::Mat;

pub use crate::MatMut;

pub use crate::MatRef;

pub use crate::Scale;

pub use crate::scale;

Macros

col

Creates a col::Col containing the arguments.

mat

Creates a Mat containing the arguments.

row

Creates a row::Row containing the arguments.

unzipped

Used to undo the zipping by the zipped! macro.

zipped

Zips together matrix of the same size, so that coefficient-wise operations can be performed on their elements.

Structs

IdentityGroup

Traits

ComplexField

Unstable trait containing the operations that a number type needs to implement.

Conjugate

Trait for types that may be implicitly conjugated.

Entity

Unstable core trait for describing how a scalar value may be split up into individual component.

RealField

Unstable trait containing the operations that a real number type needs to implement.

SimpleEntity