faer

Module perm

permutation matrix

Modules

generic

generic Perm wrapper

Structs

Own

see super::Perm

Ref

see super::PermRef

Functions

permute_cols

computes a permutation of the columns of the source matrix using the given permutation, and stores the result in the destination matrix

permute_cols_in_place

computes a permutation of the columns of the matrix using the given permutation, and stores the result in the same matrix.

permute_cols_in_place_scratch

computes the size and alignment of required workspace for applying a column permutation to a matrix in place

permute_rows

computes a permutation of the rows of the source matrix using the given permutation, and stores the result in the destination matrix

permute_rows_in_place

computes a permutation of the rows of the matrix using the given permutation, and stores the result in the same matrix

permute_rows_in_place_scratch

computes the size and alignment of required workspace for applying a row permutation to a matrix in place

swap_cols

swaps the values in the columns a and b

swap_cols_idx

swaps the two columns at indices a and b in the given matrix

swap_rows

swaps the values in the rows a and b

swap_rows_idx

swaps the two rows at indices a and b in the given matrix

Type Aliases

Perm

permutation matrix

PermRef

immutable permutation matrix view