faer::perm::generic

Struct Perm

#[repr(transparent)]
pub struct Perm<Inner>(pub Inner);

generic Perm wrapper

Tuple Fields
0: Inner
Implementations

impl<I: Index, N: Shape> Perm<Own<I, N>>

pub fn as_shape<M: Shape>(&self, dim: M) -> PermRef<'_, I, M>

returns the input permutation with the given shape after checking that it matches the current shape

pub fn into_shape<M: Shape>(self, dim: M) -> Perm<I, M>

returns the input permutation with the given shape after checking that it matches the current shape

pub fn new_checked( forward: Box<[Idx<N, I>]>, inverse: Box<[Idx<N, I>]>, dim: N, ) -> Self

creates a new permutation, by checking the validity of the inputs

panics

the function panics if any of the following conditions are violated: forward and inverse must have the same length which must be less than or equal to I::Signed::MAX, be valid permutations, and be inverse permutations of each other

pub unsafe fn new_unchecked( forward: Box<[Idx<N, I>]>, inverse: Box<[Idx<N, I>]>, ) -> Self

creates a new permutation reference, without checking the validity of the inputs

safety

forward and inverse must have the same length which must be less than or equal to I::Signed::MAX, be valid permutations, and be inverse permutations of each other

pub fn into_arrays(self) -> (Box<[Idx<N, I>]>, Box<[Idx<N, I>]>)

returns the permutation as an array

pub fn len(&self) -> N

returns the dimension of the permutation

pub fn into_inverse(self) -> Self

returns the inverse permutation

impl<I: Index, N: Shape, Inner: for<'short> Reborrow<'short, Target = Ref<'short, I, N>>> Perm<Inner>

pub fn as_ref(&self) -> PermRef<'_, I, N>

convert self to a permutation view

impl<'a, I: Index, N: Shape> Perm<Ref<'a, I, N>>

pub fn as_shape<M: Shape>(self, dim: M) -> PermRef<'a, I, M>

returns the input permutation with the given shape after checking that it matches the current shape

pub fn new_checked( forward: &'a [Idx<N, I>], inverse: &'a [Idx<N, I>], dim: N, ) -> Self

creates a new permutation, by checking the validity of the inputs

panics

the function panics if any of the following conditions are violated:

forward and inverse must have the same length which must be less than or equal to I::Signed::MAX, be valid permutations, and be inverse permutations of each other

pub unsafe fn new_unchecked( forward: &'a [Idx<N, I>], inverse: &'a [Idx<N, I>], dim: N, ) -> Self

creates a new permutation reference, without checking the validity of the inputs

safety

forward and inverse must have the same length which must be less than or equal to I::Signed::MAX, be valid permutations, and be inverse permutations of each other

pub fn arrays(self) -> (&'a [Idx<N, I>], &'a [Idx<N, I>])

returns the permutation as an array

pub fn len(&self) -> N

returns the dimension of the permutation

pub fn inverse(self) -> Self

returns the inverse permutation

pub fn canonicalized(self) -> PermRef<'a, I::FixedWidth, N>

cast the permutation to the fixed width index type

pub fn uncanonicalized<J: Index>(self) -> PermRef<'a, J, N>

cast the permutation from the fixed width index type

impl<'a, 'N, I: Index> Perm<Ref<'a, I, Dim<'N>>>

pub fn bound_arrays( self, ) -> (&'a Array<'N, Idx<Dim<'N>, I>>, &'a Array<'N, Idx<Dim<'N>, I>>)

returns the permutation as an array

impl<Inner> Perm<Inner>

pub fn from_inner_ref(inner: &Inner) -> &Self

wrap by reference

pub fn from_inner_mut(inner: &mut Inner) -> &mut Self

wrap by mutable reference