faer::perm

Struct Perm

pub struct Perm<I: Index, N: Shape = usize> { /* private fields */ }

permutation matrix

Implementations

impl<I: Index, N: Shape> Perm<I, N>

pub fn as_ref(&self) -> PermRef<'_, I, N>

convert self to a permutation view.

pub fn as_shape<M: Shape>(&self, dim: M) -> PermRef<'_, I, M>

returns the input permutation with the given shape after checking that it matches the current shape

pub fn into_shape<M: Shape>(self, dim: M) -> Perm<I, M>

returns the input permutation with the given shape after checking that it matches the current shape

pub fn new_checked( forward: Box<[Idx<N, I>]>, inverse: Box<[Idx<N, I>]>, dim: N, ) -> Self

creates a new permutation, by checking the validity of the inputs

panics

the function panics if any of the following conditions are violated: forward and inverse must have the same length which must be less than or equal to I::Signed::MAX, be valid permutations, and be inverse permutations of each other

pub unsafe fn new_unchecked( forward: Box<[Idx<N, I>]>, inverse: Box<[Idx<N, I>]>, ) -> Self

creates a new permutation reference, without checking the validity of the inputs

safety

forward and inverse must have the same length which must be less than or equal to I::Signed::MAX, be valid permutations, and be inverse permutations of each other

pub fn into_arrays(self) -> (Box<[Idx<N, I>]>, Box<[Idx<N, I>]>)

returns the permutation as an array

pub fn len(&self) -> N

returns the dimension of the permutation

pub fn into_inverse(self) -> Self

returns the inverse permutation