Module faer::prelude

faer prelude. Includes useful types and traits for solving linear systems.

Re-exports

pub use crate::col;

pub use crate::col;

pub use crate::complex_native::c32;

pub use crate::complex_native::c64;

pub use crate::linalg::solvers::Solver;

pub use crate::linalg::solvers::SolverCore;

pub use crate::linalg::solvers::SolverLstsq;

pub use crate::linalg::solvers::SolverLstsqCore;

pub use crate::linalg::solvers::SpSolver;

pub use crate::linalg::solvers::SpSolverCore;

pub use crate::linalg::solvers::SpSolverLstsq;

pub use crate::linalg::solvers::SpSolverLstsqCore;

pub use crate::mat;

pub use crate::mat;

pub use crate::row;

pub use crate::row;

pub use crate::Col;

pub use crate::ColMut;

pub use crate::ColRef;

pub use crate::Mat;

pub use crate::MatMut;

pub use crate::MatRef;

pub use crate::Row;

pub use crate::RowMut;

pub use crate::RowRef;

Macros

col

Creates a col::Col containing the arguments.

mat

Creates a Mat containing the arguments.

row

Creates a row::Row containing the arguments.

unzipped

Used to undo the zipping by the zipped! macro.

zipped

Zips together matrix of the same size, so that coefficient-wise operations can be performed on their elements.