faer
Module row

row vector

Structs

Row

heap allocated resizable row vector.

RowMut

mutable view over a row vector, similar to a mutable reference to a strided slice

RowRef

immutable view over a row vector, similar to an immutable reference to a strided slice

Traits

AsRowMut

trait for types that can be converted to a row view

AsRowRef

trait for types that can be converted to a row view

RowIndex

represents a type that can be used to slice a row, such as an index or a range of indices