faer::row

Struct RowMut

pub struct RowMut<'a, T, Cols = usize, CStride = isize> { /* private fields */ }

mutable view over a row vector, similar to a mutable reference to a strided slice

note

unlike a slice, the data pointed to by RowMut<'_, T> is allowed to be partially or fully uninitialized under certain conditions. in this case, care must be taken to not perform any operations that read the uninitialized values, or form references to them, either directly or indirectly through any of the numerical library routines, unless it is explicitly permitted

Implementations

impl<'a, T, Cols: Shape, CStride: Stride> RowMut<'a, T, Cols, CStride>

pub unsafe fn from_raw_parts_mut( ptr: *mut T, ncols: Cols, col_stride: CStride, ) -> Self

creates a RowMut from pointers to the column vector data, number of rows, and row stride

safety

this function has the same safety requirements as [MatMut::from_raw_parts_mut(ptr, 1, ncols, 0, col_stride)]

pub fn as_ptr(&self) -> *const T

returns a pointer to the row data

pub fn nrows(&self) -> usize

returns the number of rows of the row (always 1)

pub fn ncols(&self) -> Cols

returns the number of columns of the row

pub fn shape(&self) -> (usize, Cols)

returns the number of rows and columns of the row

pub fn col_stride(&self) -> CStride

returns the column stride of the row

pub fn ptr_at(&self, col: IdxInc<Cols>) -> *const T

returns a raw pointer to the element at the given index

pub unsafe fn ptr_inbounds_at(&self, col: Idx<Cols>) -> *const T

returns a raw pointer to the element at the given index, assuming the provided index is within the row bounds

safety

the behavior is undefined if any of the following conditions are violated:

col < self.ncols()

pub fn split_at_col( self, col: IdxInc<Cols>, ) -> (RowRef<'a, T, usize, CStride>, RowRef<'a, T, usize, CStride>)

see RowRef::split_at_col

pub fn transpose(self) -> ColRef<'a, T, Cols, CStride>

see RowRef::transpose

pub fn conjugate(self) -> RowRef<'a, T::Conj, Cols, CStride>
where T: Conjugate,

see RowRef::conjugate

pub fn canonical(self) -> RowRef<'a, T::Canonical, Cols, CStride>
where T: Conjugate,

see RowRef::canonical

pub fn adjoint(self) -> ColRef<'a, T::Conj, Cols, CStride>
where T: Conjugate,

see RowRef::adjoint

pub fn get<ColRange>( self, col: ColRange, ) -> <RowRef<'a, T, Cols, CStride> as RowIndex<ColRange>>::Target
where RowRef<'a, T, Cols, CStride>: RowIndex<ColRange>,

see RowRef::get

pub unsafe fn get_unchecked<ColRange>( self, col: ColRange, ) -> <RowRef<'a, T, Cols, CStride> as RowIndex<ColRange>>::Target
where RowRef<'a, T, Cols, CStride>: RowIndex<ColRange>,

see RowRef::get_unchecked

pub fn reverse_cols(self) -> RowRef<'a, T, Cols, CStride::Rev>

see RowRef::reverse_cols

pub fn subcols<V: Shape>( self, col_start: IdxInc<Cols>, ncols: V, ) -> RowRef<'a, T, V, CStride>

see RowRef::subcols

pub fn as_col_shape<V: Shape>(self, ncols: V) -> RowRef<'a, T, V, CStride>

see RowRef::as_col_shape

pub fn as_dyn_cols(self) -> RowRef<'a, T, usize, CStride>

see RowRef::as_dyn_cols

pub fn as_dyn_stride(self) -> RowRef<'a, T, Cols, isize>

see RowRef::as_dyn_stride

pub fn iter( self, ) -> impl 'a + ExactSizeIterator + DoubleEndedIterator<Item = &'a T>
where Cols: 'a,

see RowRef::iter

pub fn par_iter(self) -> impl 'a + IndexedParallelIterator<Item = &'a T>
where T: Sync, Cols: 'a,

see RowRef::par_iter

pub fn try_as_row_major(self) -> Option<RowRef<'a, T, Cols, ContiguousFwd>>

see RowRef::try_as_row_major

pub fn as_diagonal(self) -> DiagRef<'a, T, Cols, CStride>

see RowRef::as_diagonal

pub fn as_ref(&self) -> RowRef<'_, T, Cols, CStride>

returns a view over self

pub fn as_mut(&mut self) -> RowMut<'_, T, Cols, CStride>

returns a view over self

pub fn as_mat(self) -> MatRef<'a, T, usize, Cols, isize, CStride>

see RowRef::as_mat

pub fn as_mat_mut(self) -> MatMut<'a, T, usize, Cols, isize, CStride>

see RowRef::as_mat

pub fn norm_max(&self) -> Real<T>
where T: Conjugate,

see RowRef::norm_max

pub fn norm_l2(&self) -> Real<T>
where T: Conjugate,

see RowRef::norm_l2

pub fn squared_norm_l2(&self) -> Real<T>
where T: Conjugate,

see RowRef::squared_norm_l2

pub fn norm_l1(&self) -> Real<T>
where T: Conjugate,

see RowRef::norm_l1

pub fn sum(&self) -> T::Canonical
where T: Conjugate,

see RowRef::sum

pub fn cloned(&self) -> Row<T, Cols>
where T: Clone,

see RowRef::cloned

pub fn to_owned(&self) -> Row<T::Canonical, Cols>
where T: Conjugate,

see RowRef::to_owned

impl<'a, T, Cols: Shape, CStride: Stride> RowMut<'a, T, Cols, CStride>

pub fn as_ptr_mut(&self) -> *mut T

see RowRef::as_ptr

pub fn ptr_at_mut(&self, col: IdxInc<Cols>) -> *mut T

see RowRef::ptr_at

pub unsafe fn ptr_inbounds_at_mut(&self, col: Idx<Cols>) -> *mut T

see RowRef::ptr_inbounds_at

pub fn split_at_col_mut( self, col: IdxInc<Cols>, ) -> (RowMut<'a, T, usize, CStride>, RowMut<'a, T, usize, CStride>)

see RowRef::split_at_col

pub fn transpose_mut(self) -> ColMut<'a, T, Cols, CStride>

see RowRef::transpose

pub fn conjugate_mut(self) -> RowMut<'a, T::Conj, Cols, CStride>
where T: Conjugate,

see RowRef::conjugate

pub fn canonical_mut(self) -> RowMut<'a, T::Canonical, Cols, CStride>
where T: Conjugate,

see RowRef::canonical

pub fn adjoint_mut(self) -> ColMut<'a, T::Conj, Cols, CStride>
where T: Conjugate,

see RowRef::adjoint

pub fn get_mut<ColRange>( self, col: ColRange, ) -> <RowMut<'a, T, Cols, CStride> as RowIndex<ColRange>>::Target
where RowMut<'a, T, Cols, CStride>: RowIndex<ColRange>,

see RowRef::get

pub unsafe fn get_mut_unchecked<ColRange>( self, col: ColRange, ) -> <RowMut<'a, T, Cols, CStride> as RowIndex<ColRange>>::Target
where RowMut<'a, T, Cols, CStride>: RowIndex<ColRange>,

see RowRef::get

pub fn reverse_cols_mut(self) -> RowMut<'a, T, Cols, CStride::Rev>

see RowRef::reverse_cols

pub fn subcols_mut<V: Shape>( self, col_start: IdxInc<Cols>, ncols: V, ) -> RowMut<'a, T, V, CStride>

see RowRef::subcols

pub fn as_col_shape_mut<V: Shape>(self, ncols: V) -> RowMut<'a, T, V, CStride>

see RowRef::as_col_shape

pub fn as_dyn_cols_mut(self) -> RowMut<'a, T, usize, CStride>

see RowRef::as_dyn_cols

pub fn as_dyn_stride_mut(self) -> RowMut<'a, T, Cols, isize>

see RowRef::as_dyn_stride

pub fn iter_mut( self, ) -> impl 'a + ExactSizeIterator + DoubleEndedIterator<Item = &'a mut T>
where Cols: 'a,

see RowRef::iter

pub fn par_iter_mut(self) -> impl 'a + IndexedParallelIterator<Item = &'a mut T>
where T: Send, Cols: 'a,

see RowRef::par_iter

pub fn par_partition( self, count: usize, ) -> impl 'a + IndexedParallelIterator<Item = RowRef<'a, T, usize, CStride>>
where T: Sync, Cols: 'a,

see RowRef::par_partition

pub fn par_partition_mut( self, count: usize, ) -> impl 'a + IndexedParallelIterator<Item = RowMut<'a, T, usize, CStride>>
where T: Send, Cols: 'a,

see RowRef::par_partition

pub fn copy_from<RhsT: Conjugate<Canonical = T>>( &mut self, other: impl AsRowRef<T = RhsT, Cols = Cols>, )
where T: ComplexField,

copies other into self

pub fn try_as_row_major_mut(self) -> Option<RowMut<'a, T, Cols, ContiguousFwd>>

see RowRef::try_as_row_major

pub fn as_diagonal_mut(self) -> DiagMut<'a, T, Cols, CStride>

see RowRef::as_diagonal

pub fn fill(&mut self, value: T)
where T: Clone,

fills all the elements of self with value

impl<'a, T, Rows: Shape> RowMut<'a, T, Rows, ContiguousFwd>

pub fn as_slice(self) -> &'a [T]

returns a reference over the elements as a slice

impl<'a, 'ROWS, T> RowMut<'a, T, Dim<'ROWS>, ContiguousFwd>

pub fn as_array(self) -> &'a Array<'ROWS, T>

returns a reference over the elements as a lifetime-bound slice

impl<'a, T, Cols: Shape> RowMut<'a, T, Cols, ContiguousFwd>

pub fn as_slice_mut(self) -> &'a mut [T]

returns a reference over the elements as a slice

impl<'a, 'COLS, T> RowMut<'a, T, Dim<'COLS>, ContiguousFwd>

pub fn as_array_mut(self) -> &'a mut Array<'COLS, T>

returns a reference over the elements as a lifetime-bound slice