Struct faer::row::RowRef

#[repr(C)]
pub struct RowRef<'a, E: Entity> { /* private fields */ }

Immutable view over a row vector, similar to an immutable reference to a strided slice.

Note

Unlike a slice, the data pointed to by RowRef<'_, E> is allowed to be partially or fully uninitialized under certain conditions. In this case, care must be taken to not perform any operations that read the uninitialized values, or form references to them, either directly through RowRef::read, or indirectly through any of the numerical library routines, unless it is explicitly permitted.

Implementations

impl<'a, E: Entity> RowRef<'a, E>

pub fn nrows(&self) -> usize

Returns the number of rows of the row. This is always equal to 1.

pub fn ncols(&self) -> usize

Returns the number of columns of the row.

pub fn as_ptr(self) -> GroupFor<E, *const E::Unit>

Returns pointers to the matrix data.

pub fn col_stride(&self) -> isize

Returns the column stride of the matrix, specified in number of elements, not in bytes.

pub fn as_2d(self) -> MatRef<'a, E>

Returns self as a matrix view.

pub fn ptr_at(self, col: usize) -> GroupFor<E, *const E::Unit>

Returns raw pointers to the element at the given index.

pub unsafe fn ptr_inbounds_at(self, col: usize) -> GroupFor<E, *const E::Unit>

Returns raw pointers to the element at the given index, assuming the provided index is within the size of the vector.

Safety

The behavior is undefined if any of the following conditions are violated:

col < self.ncols().

pub unsafe fn split_at_unchecked(self, col: usize) -> (Self, Self)

Splits the column vector at the given index into two parts and returns an array of each subvector, in the following order:

left.

right.

Safety

The behavior is undefined if any of the following conditions are violated:

col <= self.ncols().

pub fn split_at(self, col: usize) -> (Self, Self)

Splits the column vector at the given index into two parts and returns an array of each subvector, in the following order:

top.

bottom.

Panics

The function panics if any of the following conditions are violated:

col <= self.ncols().

pub unsafe fn get_unchecked<ColRange>( self, col: ColRange, ) -> <Self as RowIndex<ColRange>>::Target
where Self: RowIndex<ColRange>,

Returns references to the element at the given index, or subvector if row is a range.

Note

The values pointed to by the references are expected to be initialized, even if the pointed-to value is not read, otherwise the behavior is undefined.

Safety

The behavior is undefined if any of the following conditions are violated:

col must be contained in [0, self.ncols()).

pub fn get<ColRange>( self, col: ColRange, ) -> <Self as RowIndex<ColRange>>::Target
where Self: RowIndex<ColRange>,

Returns references to the element at the given index, or subvector if col is a range, with bound checks.

Note

The values pointed to by the references are expected to be initialized, even if the pointed-to value is not read, otherwise the behavior is undefined.

Panics

The function panics if any of the following conditions are violated:

col must be contained in [0, self.ncols()).

pub unsafe fn read_unchecked(&self, col: usize) -> E

Reads the value of the element at the given index.

Safety

The behavior is undefined if any of the following conditions are violated:

col < self.ncols().

pub fn read(&self, col: usize) -> E

Reads the value of the element at the given index, with bound checks.

Panics

The function panics if any of the following conditions are violated:

col < self.ncols().

pub fn transpose(self) -> ColRef<'a, E>

Returns a view over the transpose of self.

pub fn conjugate(self) -> RowRef<'a, E::Conj>
where E: Conjugate,

Returns a view over the conjugate of self.

pub fn adjoint(self) -> ColRef<'a, E::Conj>
where E: Conjugate,

Returns a view over the conjugate transpose of self.

pub fn canonicalize(self) -> (RowRef<'a, E::Canonical>, Conj)
where E: Conjugate,

Returns a view over the canonical representation of self, as well as a flag declaring whether self is implicitly conjugated or not.

pub fn reverse_cols(self) -> Self

Returns a view over the self, with the columns in reversed order.

pub unsafe fn subcols_unchecked(self, col_start: usize, ncols: usize) -> Self

Returns a view over the subvector starting at column col_start, and with number of columns ncols.

Safety

The behavior is undefined if any of the following conditions are violated:

col_start <= self.ncols().

ncols <= self.ncols() - col_start.

pub fn subcols(self, col_start: usize, ncols: usize) -> Self

Returns a view over the subvector starting at col col_start, and with number of cols ncols.

Panics

The function panics if any of the following conditions are violated:

col_start <= self.ncols().

ncols <= self.ncols() - col_start.

pub fn to_owned(&self) -> Row<E::Canonical>
where E: Conjugate,

Returns an owning Row of the data.

pub fn has_nan(&self) -> bool
where E: ComplexField,

Returns true if any of the elements is NaN, otherwise returns false.

pub fn is_all_finite(&self) -> bool
where E: ComplexField,

Returns true if all of the elements are finite, otherwise returns false.

pub fn norm_max(&self) -> E::Real
where E: ComplexField,

Returns the maximum norm of self.

pub fn norm_l1(&self) -> E::Real
where E: ComplexField,

Returns the L1 norm of self.

pub fn norm_l2(&self) -> E::Real
where E: ComplexField,

Returns the L2 norm of self.

pub fn squared_norm_l2(&self) -> E::Real
where E: ComplexField,

Returns the squared L2 norm of self.

pub fn sum(&self) -> E
where E: ComplexField,

Returns the sum of self.

pub fn kron(&self, rhs: impl As2D<E>) -> Mat<E>
where E: ComplexField,

Kronecker product of self and rhs.

This is an allocating operation; see faer::linalg::kron for the allocation-free version or more info in general.

pub fn try_as_slice(self) -> Option<GroupFor<E, &'a [E::Unit]>>

Returns the row as a contiguous slice if its column stride is equal to 1.

Note

The values pointed to by the references are expected to be initialized, even if the pointed-to value is not read, otherwise the behavior is undefined.

pub fn as_ref(&self) -> RowRef<'_, E>

Returns a view over the matrix.

pub fn split_first(self) -> Option<(GroupFor<E, &'a E::Unit>, RowRef<'a, E>)>

Returns a reference to the first element and a view over the remaining ones if the row is non-empty, otherwise None.

pub fn split_last(self) -> Option<(GroupFor<E, &'a E::Unit>, RowRef<'a, E>)>

Returns a reference to the last element and a view over the remaining ones if the row is non-empty, otherwise None.

pub fn iter(self) -> ElemIter<'a, E>

Returns an iterator over the elements of the row.

pub fn chunks(self, chunk_size: usize) -> RowElemChunks<'a, E>

Returns an iterator that provides successive chunks of the elements of this row, with each having at most chunk_size elements.

pub fn partition(self, count: usize) -> RowElemPartition<'a, E>

Returns an iterator that provides exactly count successive chunks of the elements of this row.

pub fn par_chunks( self, chunk_size: usize, ) -> impl 'a + IndexedParallelIterator<Item = RowRef<'a, E>>

Returns an iterator that provides successive chunks of the elements of this row, with each having at most chunk_size elements.

Only available with the rayon feature.

pub fn par_partition( self, count: usize, ) -> impl 'a + IndexedParallelIterator<Item = RowRef<'a, E>>

Returns an iterator that provides exactly count successive chunks of the elements of this row.

Only available with the rayon feature.