faer::row

Trait RowIndex

pub trait RowIndex<ColRange> {
    type Target;

    // Required methods
    fn get(this: Self, col: ColRange) -> Self::Target;
    unsafe fn get_unchecked(this: Self, col: ColRange) -> Self::Target;
}

represents a type that can be used to slice a row, such as an index or a range of indices

Required Associated Types

type Target

sliced view type

Required Methods

fn get(this: Self, col: ColRange) -> Self::Target

slice this using col

unsafe fn get_unchecked(this: Self, col: ColRange) -> Self::Target

slice this using col without bound checks

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.