faer::sparse::csc_numeric::generic

Struct SparseColMat

#[repr(transparent)]
pub struct SparseColMat<Inner>(pub Inner);

generic SparseColMat wrapper

Tuple Fields
0: Inner
Implementations

impl<Inner> SparseColMat<Inner>

pub fn from_inner_ref(inner: &Inner) -> &Self

wrap by reference

pub fn from_inner_mut(inner: &mut Inner) -> &mut Self

wrap by mutable reference

impl<'a, Rows: Shape, Cols: Shape, I: Index, T> SparseColMat<Ref<'a, I, T, Rows, Cols>>

pub fn new( symbolic: SymbolicSparseColMatRef<'a, I, Rows, Cols>, val: &'a [T], ) -> Self

creates a new sparse matrix view.

panics

panics if the length of values is not equal to the length of symbolic.row_idx()

pub fn parts(self) -> (SymbolicSparseColMatRef<'a, I, Rows, Cols>, &'a [T])

returns the symbolic and numeric components of the sparse matrix

pub fn symbolic(&self) -> SymbolicSparseColMatRef<'a, I, Rows, Cols>

returns the symbolic component of the sparse matrix

pub fn val(self) -> &'a [T]

returns the numeric component of the sparse matrix

pub fn val_of_col(self, j: Idx<Cols>) -> &'a [T]

returns a view over the values of column j

pub fn as_shape<V: Shape, H: Shape>( self, nrows: V, ncols: H, ) -> SparseColMatRef<'a, I, T, V, H>

returns the input matrix with the given shape after checking that it matches the current shape

pub fn get(self, row: Idx<Rows>, col: Idx<Cols>) -> Option<&'a T>

returns a reference to the value at the given index, or None if the symbolic structure doesn’t contain it, or contains multiple indices with the given index

panics

panics if row >= self.nrows()

panics if col >= self.ncols()

pub fn as_dyn(self) -> SparseColMatRef<'a, I, T>

returns the input matrix with dynamic shape

pub fn as_ref(self) -> SparseColMatRef<'a, I, T, Rows, Cols>

returns a view over self

pub fn transpose(self) -> SparseRowMatRef<'a, I, T, Cols, Rows>

returns a view over the transpose of self

pub fn conjugate(self) -> SparseColMatRef<'a, I, T::Conj, Rows, Cols>
where T: Conjugate,

returns a view over the conjugate of self

pub fn adjoint(self) -> SparseRowMatRef<'a, I, T::Conj, Cols, Rows>
where T: Conjugate,

returns a view over the adjoint of self

pub fn canonical(self) -> SparseColMatRef<'a, I, T::Canonical, Rows, Cols>
where T: Conjugate,

returns an unconjugated view over self

pub fn to_row_major( &self, ) -> Result<SparseRowMat<I, T::Canonical, Rows, Cols>, FaerError>
where T: Conjugate,

returns a newly allocated matrix holding the (possibly conjugated) values of self in row major format

pub fn to_dense(&self) -> Mat<T::Canonical, Rows, Cols>
where T: Conjugate,

returns a newly allocated dense matrix holding the (possibly conjugated) values of self

pub fn triplet_iter( self, ) -> impl 'a + Iterator<Item = Triplet<Idx<Rows>, Idx<Cols>, &'a T>>
where Rows: 'a, Cols: 'a,

returns an iterator over the entries of the matrix.

impl<'a, Rows: Shape, Cols: Shape, I: Index, T> SparseColMat<Mut<'a, I, T, Rows, Cols>>

pub fn new( symbolic: SymbolicSparseColMatRef<'a, I, Rows, Cols>, val: &'a mut [T], ) -> Self

see SparseColMatRef::new

pub fn parts(self) -> (SymbolicSparseColMatRef<'a, I, Rows, Cols>, &'a [T])

see SparseColMatRef::parts

pub fn parts_mut( self, ) -> (SymbolicSparseColMatRef<'a, I, Rows, Cols>, &'a mut [T])

see SparseColMatRef::parts

pub fn symbolic(&self) -> SymbolicSparseColMatRef<'a, I, Rows, Cols>

see SparseColMatRef::symbolic

pub fn val(self) -> &'a [T]

see SparseColMatRef::val

pub fn val_mut(self) -> &'a mut [T]

see SparseColMatRef::val

pub fn val_of_col(self, j: Idx<Cols>) -> &'a [T]

see SparseColMatRef::val_of_col

pub fn val_of_col_mut(self, j: Idx<Cols>) -> &'a mut [T]

see SparseColMatRef::val_of_col

pub fn idx_val_of_col_mut( self, j: Idx<Cols>, ) -> (impl 'a + Clone + ExactSizeIterator + DoubleEndedIterator<Item = Idx<Rows>>, &'a mut [T])
where Rows: 'a, Cols: 'a,

see SymbolicSparseColMatRef::row_idx_of_col and SparseColMatRef::val_of_col

pub fn as_shape<V: Shape, H: Shape>( self, nrows: V, ncols: H, ) -> SparseColMatRef<'a, I, T, V, H>

see SparseColMatRef::as_shape

pub fn as_shape_mut<V: Shape, H: Shape>( self, nrows: V, ncols: H, ) -> SparseColMatMut<'a, I, T, V, H>

see SparseColMatRef::as_shape

pub fn as_dyn(self) -> SparseColMatRef<'a, I, T>

see SparseColMatRef::as_dyn

pub fn get(self, row: Idx<Rows>, col: Idx<Cols>) -> Option<&'a T>

see SparseColMatRef::get

pub fn get_mut(self, row: Idx<Rows>, col: Idx<Cols>) -> Option<&'a mut T>

see SparseColMatRef::get

pub fn as_dyn_mut(self) -> SparseColMatMut<'a, I, T>

see SparseColMatRef::as_dyn

pub fn transpose(self) -> SparseRowMatRef<'a, I, T, Cols, Rows>

see SparseColMatRef::transpose

pub fn transpose_mut(self) -> SparseRowMatMut<'a, I, T, Cols, Rows>

see SparseColMatRef::transpose

pub fn conjugate(self) -> SparseColMatRef<'a, I, T::Conj, Rows, Cols>
where T: Conjugate,

see SparseColMatRef::conjugate

pub fn conjugate_mut(self) -> SparseColMatMut<'a, I, T::Conj, Rows, Cols>
where T: Conjugate,

see SparseColMatRef::conjugate

pub fn adjoint(self) -> SparseRowMatRef<'a, I, T::Conj, Cols, Rows>
where T: Conjugate,

see SparseColMatRef::adjoint

pub fn adjoint_mut(self) -> SparseRowMatMut<'a, I, T::Conj, Cols, Rows>
where T: Conjugate,

see SparseColMatRef::adjoint

pub fn canonical(self) -> SparseColMatRef<'a, I, T::Canonical, Rows, Cols>
where T: Conjugate,

see SparseColMatRef::canonical

pub fn canonical_mut(self) -> SparseColMatMut<'a, I, T::Canonical, Rows, Cols>
where T: Conjugate,

see SparseColMatRef::canonical

pub fn to_row_major( &self, ) -> Result<SparseRowMat<I, T::Canonical, Rows, Cols>, FaerError>
where T: Conjugate,

see SparseColMatRef::to_row_major

pub fn to_dense(&self) -> Mat<T::Canonical, Rows, Cols>
where T: Conjugate,

see SparseColMatRef::to_dense

pub fn triplet_iter( self, ) -> impl 'a + Iterator<Item = Triplet<Idx<Rows>, Idx<Cols>, &'a T>>
where Rows: 'a, Cols: 'a,

see SparseColMatRef::triplet_iter

impl<Rows: Shape, Cols: Shape, I: Index, T> SparseColMat<Own<I, T, Rows, Cols>>

pub fn new(symbolic: SymbolicSparseColMat<I, Rows, Cols>, val: Vec<T>) -> Self

see SparseColMatRef::new

pub fn parts(&self) -> (SymbolicSparseColMatRef<'_, I, Rows, Cols>, &[T])

see SparseColMatRef::parts

pub fn parts_mut( &mut self, ) -> (SymbolicSparseColMatRef<'_, I, Rows, Cols>, &mut [T])

see SparseColMatRef::parts

pub fn into_parts(self) -> (SymbolicSparseColMat<I, Rows, Cols>, Vec<T>)

see SparseColMatRef::parts

pub fn symbolic(&self) -> SymbolicSparseColMatRef<'_, I, Rows, Cols>

see SparseColMatRef::symbolic

pub fn val(&self) -> &[T]

see SparseColMatRef::val

pub fn val_mut(&mut self) -> &mut [T]

see SparseColMatRef::val

pub fn val_of_col(&self, j: Idx<Cols>) -> &[T]

see SparseColMatRef::val_of_col

pub fn val_of_col_mut(&mut self, j: Idx<Cols>) -> &mut [T]

see SparseColMatRef::val_of_col

pub fn as_shape<V: Shape, H: Shape>( &self, nrows: V, ncols: H, ) -> SparseColMatRef<'_, I, T, V, H>

see SparseColMatRef::as_shape

pub fn as_shape_mut<V: Shape, H: Shape>( &mut self, nrows: V, ncols: H, ) -> SparseColMatMut<'_, I, T, V, H>

see SparseColMatRef::as_shape

pub fn into_shape<V: Shape, H: Shape>( self, nrows: V, ncols: H, ) -> SparseColMat<I, T, V, H>

see SparseColMatRef::as_shape

pub fn as_dyn(&self) -> SparseColMatRef<'_, I, T>

see SparseColMatRef::as_dyn

pub fn as_dyn_mut(&mut self) -> SparseColMatMut<'_, I, T>

see SparseColMatRef::as_dyn

pub fn into_dyn(self) -> SparseColMat<I, T>

see SparseColMatRef::as_dyn

pub fn as_ref(&self) -> SparseColMatRef<'_, I, T, Rows, Cols>

see SparseColMatRef::as_ref

pub fn transpose(&self) -> SparseRowMatRef<'_, I, T, Cols, Rows>

see SparseColMatRef::transpose

pub fn transpose_mut(&mut self) -> SparseRowMatMut<'_, I, T, Cols, Rows>

see SparseColMatRef::transpose

pub fn into_transpose(self) -> SparseRowMat<I, T, Cols, Rows>

see SparseColMatRef::transpose

pub fn conjugate(&self) -> SparseColMatRef<'_, I, T::Conj, Rows, Cols>
where T: Conjugate,

see SparseColMatRef::conjugate

pub fn conjugate_mut(&mut self) -> SparseColMatMut<'_, I, T::Conj, Rows, Cols>
where T: Conjugate,

see SparseColMatRef::conjugate

pub fn into_conjugate(self) -> SparseColMat<I, T::Conj, Rows, Cols>
where T: Conjugate,

see SparseColMatRef::conjugate

pub fn adjoint(&self) -> SparseRowMatRef<'_, I, T::Conj, Cols, Rows>
where T: Conjugate,

see SparseColMatRef::adjoint

pub fn adjoint_mut(&mut self) -> SparseRowMatMut<'_, I, T::Conj, Cols, Rows>
where T: Conjugate,

see SparseColMatRef::adjoint

pub fn into_adjoint(self) -> SparseRowMat<I, T::Conj, Cols, Rows>
where T: Conjugate,

see SparseColMatRef::adjoint

pub fn canonical(&self) -> SparseColMatRef<'_, I, T::Canonical, Rows, Cols>
where T: Conjugate,

see SparseColMatRef::canonical

pub fn canonical_mut( &mut self, ) -> SparseColMatMut<'_, I, T::Canonical, Rows, Cols>
where T: Conjugate,

see SparseColMatRef::canonical

pub fn into_canonical(self) -> SparseColMat<I, T::Canonical, Rows, Cols>
where T: Conjugate,

see SparseColMatRef::canonical

pub fn to_row_major( &self, ) -> Result<SparseRowMat<I, T::Canonical, Rows, Cols>, FaerError>
where T: Conjugate,

see SparseColMatRef::to_row_major

pub fn to_dense(&self) -> Mat<T::Canonical, Rows, Cols>
where T: Conjugate,

see SparseColMatRef::to_dense

pub fn new_from_argsort( symbolic: SymbolicSparseColMat<I, Rows, Cols>, argsort: &Argsort<I>, val: &[T], ) -> Result<Self, FaerError>
where T: ComplexField,

create a new matrix from a previously created symbolic structure and value order

the provided values must correspond to the same indices that were provided in the function call from which the order was created

pub fn try_new_from_triplets( nrows: Rows, ncols: Cols, entries: &[Triplet<Idx<Rows, I>, Idx<Cols, I>, T>], ) -> Result<Self, CreationError>
where T: ComplexField,

create a new matrix from triplets

pub fn try_new_from_nonnegative_triplets( nrows: Rows, ncols: Cols, entries: &[Triplet<MaybeIdx<Rows, I>, MaybeIdx<Cols, I>, T>], ) -> Result<Self, CreationError>
where T: ComplexField,

create a new matrix from triplets

negative indices are ignored

pub fn get(&self, row: Idx<Rows>, col: Idx<Cols>) -> Option<&T>

see SparseColMatRef::get

pub fn get_mut(&mut self, row: Idx<Rows>, col: Idx<Cols>) -> Option<&mut T>

see SparseColMatRef::get

pub fn triplet_iter( &self, ) -> impl '_ + Iterator<Item = Triplet<Idx<Rows>, Idx<Cols>, &T>>

see SparseColMatRef::triplet_iter

impl<I: Index, T: ComplexField, Inner: for<'short> Reborrow<'short, Target = Ref<'short, I, T>>> SparseColMat<Inner>

pub fn sp_solve_lower_triangular_in_place( &self, rhs: impl AsMatMut<T = T, Rows = usize>, )

assuming self is a lower triangular matrix, solves the equation self * x = rhs, and stores the result in rhs

note

the matrix indices need not be sorted, but the diagonal element is assumed to be the first stored element in each column

pub fn sp_solve_upper_triangular_in_place( &self, rhs: impl AsMatMut<T = T, Rows = usize>, )

assuming self is an upper triangular matrix, solves the equation self * x = rhs, and stores the result in rhs

note

the matrix indices need not be sorted, but the diagonal element is assumed to be the last stored element in each column

pub fn sp_solve_unit_lower_triangular_in_place( &self, rhs: impl AsMatMut<T = T, Rows = usize>, )

assuming self is a unit lower triangular matrix, solves the equation self * x = rhs, and stores the result in rhs

note

the matrix indices need not be sorted, but the diagonal element is assumed to be the first stored element in each column

pub fn sp_solve_unit_upper_triangular_in_place( &self, rhs: impl AsMatMut<T = T, Rows = usize>, )

assuming self is a unit upper triangular matrix, solves the equation self * x = rhs, and stores the result in rhs

note

the matrix indices need not be sorted, but the diagonal element is assumed to be the last stored element in each column

pub fn sp_cholesky(&self, side: Side) -> Result<Llt<I, T>, LltError>

returns the $LL^\top$ decomposition of self. only the provided side is accessed

pub fn sp_lu(&self) -> Result<Lu<I, T>, LuError>

returns the $LU$ decomposition of self with partial (row) pivoting

pub fn sp_qr(&self) -> Result<Qr<I, T>, FaerError>

returns the $QR$ decomposition of self