faer::sparse::linalg::cholesky

Module simplicial

simplicial factorization module

a simplicial factorization is one that processes the elements of the cholesky factor of the input matrix single elements, rather than by blocks. this is more efficient if the cholesky factor is very sparse

Structs

EliminationTreeRef

reference to a slice containing the cholesky factor’s elimination tree

SimplicialLdltRef

cholesky $LDL^H$ factors containing both the symbolic and numeric representations

SimplicialLltRef

cholesky $LL^H$ factor containing both its symbolic and numeric representations

SymbolicSimplicialCholesky

cholesky factor structure containing its symbolic structure

Functions

factorize_simplicial_numeric_ldlt

computes the numeric values of the cholesky $LDL^H$ factors of the matrix $A$, and stores them in l_values

factorize_simplicial_numeric_ldlt_scratch

returns the size and alignment of the workspace required to compute the numeric cholesky $LDL^H$ factorization of a matrix $A$ with dimension n

factorize_simplicial_numeric_ldlt_with_row_idx

computes the row indices and numeric values of the cholesky $LDL^H$ factor of the matrix $A$, and stores them in l_row_idx and l_values

factorize_simplicial_numeric_llt

computes the numeric values of the cholesky $LL^H$ factor of the matrix $A$, and stores them in l_values

factorize_simplicial_numeric_llt_scratch

returns the size and alignment of the workspace required to compute the numeric cholesky $LL^H$ factorization of a matrix $A$ with dimension n

factorize_simplicial_numeric_llt_with_row_idx

computes the row indices and numeric values of the cholesky $LL^H$ factor of the matrix $A$, and stores them in l_row_idx and l_values

factorize_simplicial_symbolic_cholesky

computes the symbolic structure of the cholesky factor of the matrix $A$

factorize_simplicial_symbolic_cholesky_scratch

computes the size and alignment of the workspace required to compute the symbolic cholesky factorization of a square matrix with size n

prefactorize_symbolic_cholesky

computes the elimination tree and column counts of the cholesky factorization of the matrix $A$

prefactorize_symbolic_cholesky_scratch

computes the size and alignment of the workspace required to compute the elimination tree and column counts of a matrix of size n with nnz non-zero entries