faer::sparse::linalg::cholesky::simplicial

Function factorize_simplicial_numeric_ldlt_with_row_idx

pub fn factorize_simplicial_numeric_ldlt_with_row_idx<I: Index, T: ComplexField>(
    L_values: &mut [T],
    L_row_idx: &mut [I],
    L_col_ptr: &[I],
    etree: EliminationTreeRef<'_, I>,
    A: SparseColMatRef<'_, I, T>,
    regularization: LdltRegularization<'_, T::Real>,
    stack: &mut MemStack,
) -> Result<LdltInfo, LdltError>

computes the row indices and numeric values of the cholesky $LDL^H$ factor of the matrix $A$, and stores them in l_row_idx and l_values

note

only the upper triangular part of $A$ is accessed

panics

the elimination tree and column counts must be computed by calling prefactorize_symbolic_cholesky with the same matrix, then the column pointers are computed from a prefix sum of the column counts. otherwise, the behavior is unspecified and panics may occur