Function faer::sparse::linalg::cholesky::simplicial::factorize_simplicial_numeric_llt_with_row_indices

pub fn factorize_simplicial_numeric_llt_with_row_indices<I: Index, E: ComplexField>(
    L_values: GroupFor<E, &mut [E::Unit]>,
    L_row_indices: &mut [I],
    L_col_ptrs: &[I],
    etree: EliminationTreeRef<'_, I>,
    A: SparseColMatRef<'_, I, E>,
    regularization: LltRegularization<E>,
    stack: PodStack<'_>,
) -> Result<usize, CholeskyError>

Computes the row indices and numeric values of the Cholesky LLT factor of the matrix A, and stores them in L_row_indices and L_values.

Note

Only the upper triangular part of A is accessed.

Panics

The elimination tree and column counts must be computed by calling prefactorize_symbolic_cholesky with the same matrix, then the column pointers are computed from a prefix sum of the column counts. Otherwise, the behavior is unspecified and panics may occur.