faer::sparse::linalg::cholesky
Module supernodal

supernodal factorization module

a supernodal factorization is one that processes the elements of the cholesky factor of the input matrix by blocks, rather than single elements. this is more efficient if the cholesky factor is somewhat dense

Structs

SupernodalIntranodeBunchKaufmanRef

cholesky bunch-kaufman factors containing both the symbolic and numeric representations

SupernodalLdltRef

cholesky $LDL^H$ factors containing both the symbolic and numeric representations

SupernodalLltRef

cholesky $LL^H$ factor containing both its symbolic and numeric representations

SupernodeRef

a single supernode from the cholesky factor

SymbolicSupernodalCholesky

cholesky factor structure containing its symbolic structure

SymbolicSupernodeRef

symbolic structure of a single supernode from the cholesky factor

Functions

factorize_supernodal_numeric_intranode_bunch_kaufman

computes the numeric values of the cholesky bunch-kaufman factors of the matrix $A$ with intranodal pivoting, and stores them in l_values

factorize_supernodal_numeric_intranode_bunch_kaufman_scratch

returns the size and alignment of the workspace required to compute the numeric cholesky bunch-kaufman factorization with intranodal pivoting of a matrix $A$ with dimension n

factorize_supernodal_numeric_ldlt

computes the numeric values of the cholesky $LDL^H$ factors of the matrix $A$, and stores them in l_values

factorize_supernodal_numeric_ldlt_scratch

returns the size and alignment of the workspace required to compute the numeric cholesky $LDL^H$ factorization of a matrix $A$ with dimension n

factorize_supernodal_numeric_llt

computes the numeric values of the cholesky $LL^H$ factor of the matrix $A$, and stores them in l_values

factorize_supernodal_numeric_llt_scratch

returns the size and alignment of the workspace required to compute the numeric cholesky $LL^H$ factorization of a matrix $A$ with dimension n

factorize_supernodal_symbolic_cholesky

computes the supernodal symbolic structure of the cholesky factor of the matrix $A$

factorize_supernodal_symbolic_cholesky_scratch

returns the size and alignment of the workspace required to compute the symbolic supernodal factorization of a matrix of size n