faer::sparse::linalg::cholesky::supernodal

Function factorize_supernodal_numeric_llt

pub fn factorize_supernodal_numeric_llt<I: Index, T: ComplexField>(
    L_values: &mut [T],
    A_lower: SparseColMatRef<'_, I, T>,
    regularization: LltRegularization<T::Real>,
    symbolic: &SymbolicSupernodalCholesky<I>,
    par: Par,
    stack: &mut MemStack,
    params: Spec<LltParams, T>,
) -> Result<LltInfo, LltError>

computes the numeric values of the cholesky $LL^H$ factor of the matrix $A$, and stores them in l_values

warning

only the lower (not upper, unlike the other functions) triangular part of $A$ is accessed

panics

the symbolic structure must be computed by calling factorize_supernodal_symbolic_cholesky on a matrix with the same symbolic structure otherwise, the behavior is unspecified and panics may occur