faer::sparse::linalg

Module lu

computes the $LU$ decomposition of a given sparse matrix. see faer::linalg::lu for more info

the entry point in this module is SymbolicLu and factorize_symbolic_lu

note

the functions in this module accept unsorted inputs, and may produce unsorted decomposition factors.

Modules

simplicial

simplicial factorization module

supernodal

supernodal factorization module

Structs

LuRef

sparse $LU$ factorization wrapper

LuSymbolicParams

tuning parameters for the $LU$ symbolic factorization

NumericLu

structure that contains the numerical values and row pivoting permutation of the lu decomposition

SymbolicLu

the symbolic structure of a sparse $LU$ decomposition

Enums

SymbolicLuRaw

the inner factorization used for the symbolic $LU$, either simplicial or symbolic

Functions

factorize_symbolic_lu

computes the symbolic $LU$ factorization of the matrix $A$, or returns an error if the operation could not be completed