faer::sparse::linalg::lu

Module simplicial

simplicial factorization module

a supernodal factorization is one that processes the elements of the $LU$ factors of the input matrix by single elements, rather than by blocks. this is more efficient if the lu factors are very sparse

Structs

SimplicialLu

$LU$ factor structure containing the symbolic and numerical representations

Functions

factorize_simplicial_numeric_lu

computes the numeric values of the $LU$ factors of the matrix $A$ as well as the row pivoting permutation, and stores them in lu and row_perm/row_perm_inv

factorize_simplicial_numeric_lu_scratch

computes the size and alignment of the workspace required to perform a numeric $LU$ factorization