faer::sparse::linalg::lu

Module supernodal

supernodal factorization module

a supernodal factorization is one that processes the elements of the $LU$ factors of the input matrix by blocks, rather than by single elements. this is more efficient if the lu factors are somewhat dense

Structs

SupernodalLu

$LU$ factor structure containing the symbolic and numerical representations

SymbolicSupernodalLu

$LU$ factor structure containing the symbolic structure

Functions

factorize_supernodal_numeric_lu

computes the numeric values of the $LU$ factors of the matrix $A$ as well as the row pivoting permutation, and stores them in lu and row_perm/row_perm_inv

factorize_supernodal_numeric_lu_scratch

computes the size and alignment of the workspace required to perform a numeric $LU$ factorization

factorize_supernodal_symbolic_lu

computes the symbolic structure of the $LU$ factors of the matrix $A$

factorize_supernodal_symbolic_lu_scratch

computes the size and alignment of the workspace required to compute the symbolic $LU$ factorization of a square matrix with size n.