faer::sparse::linalg::matmul

Function dense_sparse_matmul

pub fn dense_sparse_matmul<I: Index, T: ComplexField, LhsT: Conjugate<Canonical = T>, RhsT: Conjugate<Canonical = T>>(
    dst: MatMut<'_, T>,
    beta: Accum,
    lhs: MatRef<'_, LhsT>,
    rhs: SparseColMatRef<'_, I, RhsT>,
    alpha: T,
    par: Par,
)

multiplies a dense matrix lhs by a sparse matrix rhs, and stores or adds the result to dst. see faer::linalg::matmul::matmul for more details.

note

allows unsorted matrices.