faer::sparse::linalg::matmul

Function sparse_sparse_matmul

pub fn sparse_sparse_matmul<I: Index, T: ComplexField, LhsT: Conjugate<Canonical = T>, RhsT: Conjugate<Canonical = T>>(
    lhs: SparseColMatRef<'_, I, LhsT>,
    rhs: SparseColMatRef<'_, I, RhsT>,
    alpha: T,
    par: Par,
) -> Result<SparseColMat<I, T>, FaerError>

performs a numeric matrix multiplication of a sparse matrix lhs by a sparse matrix rhs multiplied by alpha, and returns the result.

note

lhs and rhs are allowed to be unsorted matrices.