faer::sparse::linalg::matmul

Function sparse_sparse_matmul_numeric

pub fn sparse_sparse_matmul_numeric<I: Index, T: ComplexField, LhsT: Conjugate<Canonical = T>, RhsT: Conjugate<Canonical = T>>(
    dst: SparseColMatMut<'_, I, T>,
    beta: Accum,
    lhs: SparseColMatRef<'_, I, LhsT>,
    rhs: SparseColMatRef<'_, I, RhsT>,
    alpha: T,
    info: &SparseMatMulInfo,
    par: Par,
    stack: &mut MemStack,
)

performs a numeric matrix multiplication of a sparse matrix lhs by a sparse matrix rhs multiplied by alpha, and stores or adds the result to dst.

note

lhs and rhs are allowed to be unsorted matrices.