faer::sparse::linalg

Module qr

computes the $QR$ decomposition of a given sparse matrix. see crate::linalg::qr for more info

the entry point in this module is SymbolicQr and factorize_symbolic_qr

note

the functions in this module accept unsorted inputs, and may produce unsorted decomposition factors.

Modules

simplicial

simplicial factorization module

supernodal

supernodal factorization module

Structs

QrRef

sparse $QR$ factorization wrapper

QrSymbolicParams

tuning parameters for the $QR$ symbolic factorization

SymbolicQr

the symbolic structure of a sparse $QR$ decomposition

Enums

SymbolicQrRaw

the inner factorization used for the symbolic $QR$, either simplicial or symbolic

Functions

col_etree

computes the column elimination tree of $A$, which is the same as the elimination tree of $A^\top A$

col_etree_scratch

computes the size and alignment of the workspace required to compute the column elimination tree of a matrix $A$ with dimensions (nrows, ncols)

column_counts_aat_scratch

computes the size and alignment of the workspace required to compute the column counts of the cholesky factor of the matrix $A A^\top$, where $A$ has dimensions (nrows, ncols)

column_counts_ata

computes the column counts of the cholesky factor of $A^\top A$

factorize_symbolic_qr

computes the symbolic $QR$ factorization of the matrix $A$, or returns an error if the operation could not be completed

postorder

computes a postordering of the elimination tree of size n

postorder_scratch

computes the size and alignment of the workspace required to compute the postordering of an elimination tree of size n