Trait faer::sparse::linalg::solvers::SpSolver

pub trait SpSolver<E: ComplexField>: SpSolverCore<E> {
    // Required methods
    fn solve_in_place(&self, rhs: impl ColBatchMut<E>);
    fn solve_conj_in_place(&self, rhs: impl ColBatchMut<E>);
    fn solve_transpose_in_place(&self, rhs: impl ColBatchMut<E>);
    fn solve_conj_transpose_in_place(&self, rhs: impl ColBatchMut<E>);
    fn solve<ViewE: Conjugate<Canonical = E>, B: ColBatch<ViewE>>(
        &self,
        rhs: B,
    ) -> B::Owned;
    fn solve_conj<ViewE: Conjugate<Canonical = E>, B: ColBatch<ViewE>>(
        &self,
        rhs: B,
    ) -> B::Owned;
    fn solve_transpose<ViewE: Conjugate<Canonical = E>, B: ColBatch<ViewE>>(
        &self,
        rhs: B,
    ) -> B::Owned;
    fn solve_conj_transpose<ViewE: Conjugate<Canonical = E>, B: ColBatch<ViewE>>(
        &self,
        rhs: B,
    ) -> B::Owned;
}

Solver that can compute solution of a linear system.

Required Methods

fn solve_in_place(&self, rhs: impl ColBatchMut<E>)

Solves the equation self * X = rhs when self is square, and stores the result in rhs.

fn solve_conj_in_place(&self, rhs: impl ColBatchMut<E>)

Solves the equation conjugate(self) * X = rhs when self is square, and stores the result in rhs.

fn solve_transpose_in_place(&self, rhs: impl ColBatchMut<E>)

Solves the equation transpose(self) * X = rhs when self is square, and stores the result in rhs.

fn solve_conj_transpose_in_place(&self, rhs: impl ColBatchMut<E>)

Solves the equation adjoint(self) * X = rhs when self is square, and stores the result in rhs.

fn solve<ViewE: Conjugate<Canonical = E>, B: ColBatch<ViewE>>( &self, rhs: B, ) -> B::Owned

Solves the equation self * X = rhs when self is square, and returns the result.

fn solve_conj<ViewE: Conjugate<Canonical = E>, B: ColBatch<ViewE>>( &self, rhs: B, ) -> B::Owned

Solves the equation conjugate(self) * X = rhs when self is square, and returns the result.

fn solve_transpose<ViewE: Conjugate<Canonical = E>, B: ColBatch<ViewE>>( &self, rhs: B, ) -> B::Owned

Solves the equation transpose(self) * X = rhs when self is square, and returns the result.

fn solve_conj_transpose<ViewE: Conjugate<Canonical = E>, B: ColBatch<ViewE>>( &self, rhs: B, ) -> B::Owned

Solves the equation adjoint(self) * X = rhs when self is square, and returns the result.

Object Safety

This trait is not object safe.