faer::sparse::linalg::triangular_solve

Function solve_upper_triangular_transpose_in_place

pub fn solve_upper_triangular_transpose_in_place<I: Index, T: ComplexField>(
    triu: SparseColMatRef<'_, I, T>,
    conj_triu: Conj,
    rhs: MatMut<'_, T>,
    par: Par,
)

assuming triu is an upper triangular matrix, solves the equation triu.transpose() * x = rhs, and stores the result in rhs, implicitly conjugating triu if needed

note

the matrix indices need not be sorted, but the diagonal element is assumed to be the first stored element in each column.