faer::sparse
Module ops

sparse matrix binary and ternary operation implementations

Functions

add

returns the sum of lhs and rhs.

add_assign

computes the sum of dst and src and stores the result in dst without changing its symbolic structure.

add_into

computes the sum of lhs and rhs, storing the result in dst without changing its symbolic structure.

binary_op

returns the resulting matrix obtained by applying f to the elements from lhs and rhs, skipping entries that are unavailable in both of lhs and rhs.

binary_op_assign_into

returns the resulting matrix obtained by applying f to the elements from dst and src skipping entries that are unavailable in both of them.
the sparsity patter of dst is unchanged.

sub

returns the difference of lhs and rhs.

sub_assign

computes the difference of dst and src and stores the result in dst without changing its symbolic structure.

sub_into

computes the difference of lhs and rhs, storing the result in dst without changing its symbolic structure.

ternary_op_assign_into

returns the resulting matrix obtained by applying f to the elements from dst, lhs and rhs, skipping entries that are unavailable in all of dst, lhs and rhs.
the sparsity patter of dst is unchanged.

union_symbolic

returns the sparsity pattern containing the union of those of lhs and rhs.