faer::sparse::ops

Function sub_into

pub fn sub_into<I: Index, T: ComplexField, LhsT: Conjugate<Canonical = T>, RhsT: Conjugate<Canonical = T>>(
    dst: SparseColMatMut<'_, I, T>,
    lhs: SparseColMatRef<'_, I, LhsT>,
    rhs: SparseColMatRef<'_, I, RhsT>,
)

computes the difference of lhs and rhs, storing the result in dst without changing its symbolic structure.

panics

panics if dst, lhs and rhs don’t have matching dimensions.
panics if lhs or rhs contains an index that’s unavailable in dst.