faer::sparse
Struct SparseColMatRef

pub struct SparseColMatRef<'a, I, T, Rows = usize, Cols = usize> { /* private fields */ }

view over sparse column major matrix

see SymbolicSparseColMatRef

Implementations

impl<'a, Rows: Shape, Cols: Shape, I: Index, T> SparseColMatRef<'a, I, T, Rows, Cols>

pub fn new( symbolic: SymbolicSparseColMatRef<'a, I, Rows, Cols>, val: &'a [T], ) -> Self

creates a new sparse matrix view.

panics

panics if the length of values is not equal to the length of symbolic.row_idx()

pub fn parts(self) -> (SymbolicSparseColMatRef<'a, I, Rows, Cols>, &'a [T])

returns the symbolic and numeric components of the sparse matrix

pub fn symbolic(&self) -> SymbolicSparseColMatRef<'a, I, Rows, Cols>

returns the symbolic component of the sparse matrix

pub fn val(self) -> &'a [T]

returns the numeric component of the sparse matrix

pub fn val_of_col(self, j: Idx<Cols>) -> &'a [T]

returns a view over the values of column j

pub fn as_shape<V: Shape, H: Shape>( self, nrows: V, ncols: H, ) -> SparseColMatRef<'a, I, T, V, H>

returns the input matrix with the given shape after checking that it matches the current shape

pub fn get(self, row: Idx<Rows>, col: Idx<Cols>) -> Option<&'a T>

returns a reference to the value at the given index, or None if the symbolic structure doesnt contain it, or contains multiple indices with the given index

panics

panics if row >= self.nrows()

panics if col >= self.ncols()

pub fn as_dyn(self) -> SparseColMatRef<'a, I, T>

returns the input matrix with dynamic shape

pub fn as_ref(self) -> SparseColMatRef<'a, I, T, Rows, Cols>

returns a view over self

pub fn transpose(self) -> SparseRowMatRef<'a, I, T, Cols, Rows>

returns a view over the transpose of self

pub fn conjugate(self) -> SparseColMatRef<'a, I, T::Conj, Rows, Cols>
where T: Conjugate,

returns a view over the conjugate of self

pub fn adjoint(self) -> SparseRowMatRef<'a, I, T::Conj, Cols, Rows>
where T: Conjugate,

returns a view over the adjoint of self

pub fn canonical(self) -> SparseColMatRef<'a, I, T::Canonical, Rows, Cols>
where T: Conjugate,

returns an unconjugated view over self

pub fn to_row_major( &self, ) -> Result<SparseRowMat<I, T::Canonical, Rows, Cols>, FaerError>
where T: Conjugate,

returns a newly allocated matrix holding the (possibly conjugated) values of self in row major format

pub fn to_dense(&self) -> Mat<T::Canonical, Rows, Cols>
where T: Conjugate,

returns a newly allocated dense matrix holding the (possibly conjugated) values of self

impl<I: Index, T: ComplexField> SparseColMatRef<'_, I, T>

pub fn sp_solve_lower_triangular_in_place( &self, rhs: impl AsMatMut<T = T, Rows = usize>, )

assuming self is a lower triangular matrix, solves the equation self * x = rhs, and stores the result in rhs

note

the matrix indices need not be sorted, but the diagonal element is assumed to be the first stored element in each column

pub fn sp_solve_upper_triangular_in_place( &self, rhs: impl AsMatMut<T = T, Rows = usize>, )

assuming self is an upper triangular matrix, solves the equation self * x = rhs, and stores the result in rhs

note

the matrix indices need not be sorted, but the diagonal element is assumed to be the last stored element in each column

pub fn sp_solve_unit_lower_triangular_in_place( &self, rhs: impl AsMatMut<T = T, Rows = usize>, )

assuming self is a unit lower triangular matrix, solves the equation self * x = rhs, and stores the result in rhs

note

the matrix indices need not be sorted, but the diagonal element is assumed to be the first stored element in each column

pub fn sp_solve_unit_upper_triangular_in_place( &self, rhs: impl AsMatMut<T = T, Rows = usize>, )

assuming self is a unit upper triangular matrix, solves the equation self * x = rhs, and stores the result in rhs

note

the matrix indices need not be sorted, but the diagonal element is assumed to be the last stored element in each column

pub fn sp_cholesky(&self, side: Side) -> Result<Llt<I, T>, LltError>

returns the $LL^\top$ decomposition of self. only the provided side is accessed

pub fn sp_lu(&self) -> Result<Lu<I, T>, LuError>

returns the $LU$ decomposition of self with partial (row) pivoting

pub fn sp_qr(&self) -> Result<Qr<I, T>, FaerError>

returns the $QR$ decomposition of self