Struct faer::sparse::SymbolicSparseColMat

pub struct SymbolicSparseColMat<I: Index> { /* private fields */ }

Symbolic structure of sparse matrix in column format, either compressed or uncompressed.

Requires:

nrows <= I::Signed::MAX (always checked)

ncols <= I::Signed::MAX (always checked)

col_ptrs has length ncols + 1 (always checked)

col_ptrs is non-decreasing

col_ptrs[0]..col_ptrs[ncols] is a valid range in row_indices (always checked, assuming non-decreasing)

if nnz_per_col is None, elements of row_indices[col_ptrs[j]..col_ptrs[j + 1]] are less than nrows

nnz_per_col[j] <= col_ptrs[j+1] - col_ptrs[j]

if nnz_per_col is Some(_), elements of row_indices[col_ptrs[j]..][..nnz_per_col[j]] are less than nrows

Within each column, row indices are sorted in non-decreasing order.

Note

Some algorithms allow working with matrices containing unsorted row indices per column.

Passing such a matrix to an algorithm that does not explicitly permit this is unspecified (though not undefined) behavior.

Implementations

impl<I: Index> SymbolicSparseColMat<I>

pub fn new_checked( nrows: usize, ncols: usize, col_ptrs: Vec<I>, nnz_per_col: Option<Vec<I>>, row_indices: Vec<I>, ) -> Self

Creates a new symbolic matrix view after asserting its invariants.

Panics

See type level documentation.

pub fn new_unsorted_checked( nrows: usize, ncols: usize, col_ptrs: Vec<I>, nnz_per_col: Option<Vec<I>>, row_indices: Vec<I>, ) -> Self

Creates a new symbolic matrix view from data containing duplicate and/or unsorted row indices per column, after asserting its other invariants.

Panics

See type level documentation.

pub unsafe fn new_unchecked( nrows: usize, ncols: usize, col_ptrs: Vec<I>, nnz_per_col: Option<Vec<I>>, row_indices: Vec<I>, ) -> Self

Creates a new symbolic matrix view without asserting its invariants.

Safety

See type level documentation.

pub fn into_parts(self) -> (usize, usize, Vec<I>, Option<Vec<I>>, Vec<I>)

Returns the components of the matrix in the order:

row count,

column count,

column pointers,

nonzeros per column,

row indices.

pub fn as_ref(&self) -> SymbolicSparseColMatRef<'_, I>

Returns a view over the symbolic structure of self.

pub fn nrows(&self) -> usize

Returns the number of rows of the matrix.

pub fn ncols(&self) -> usize

Returns the number of columns of the matrix.

pub fn shape(&self) -> (usize, usize)

Returns the number of rows and columns of the matrix.

pub fn into_transpose(self) -> SymbolicSparseRowMat<I>

Consumes the matrix, and returns its transpose in row-major format without reallocating.

Note

Allows unsorted matrices, producing an unsorted output.

pub fn to_owned(&self) -> Result<SymbolicSparseColMat<I>, FaerError>

Copies the current matrix into a newly allocated matrix.

Note

Allows unsorted matrices, producing an unsorted output.

pub fn to_row_major(&self) -> Result<SymbolicSparseRowMat<I>, FaerError>

Copies the current matrix into a newly allocated matrix, with row-major order.

Note

Allows unsorted matrices, producing a sorted output. Duplicate entries are kept, however.

pub fn compute_nnz(&self) -> usize

Returns the number of symbolic non-zeros in the matrix.

The value is guaranteed to be less than I::Signed::MAX.

Note

Allows unsorted matrices, but the output is a count of all the entries, including the duplicate ones.

pub fn col_ptrs(&self) -> &[I]

Returns the column pointers.

pub fn nnz_per_col(&self) -> Option<&[I]>

Returns the count of non-zeros per column of the matrix.

pub fn row_indices(&self) -> &[I]

Returns the row indices.

pub fn row_indices_of_col_raw(&self, j: usize) -> &[I]

Returns the row indices of column j.

Panics

Panics if j >= self.ncols().

pub fn row_indices_of_col( &self, j: usize, ) -> impl '_ + ExactSizeIterator + DoubleEndedIterator<Item = usize>

Returns the row indices of column j.

Panics

Panics if j >= self.ncols().

pub fn col_range(&self, j: usize) -> Range<usize>

Returns the range that the column j occupies in self.row_indices().

Panics

Panics if j >= self.ncols().

pub unsafe fn col_range_unchecked(&self, j: usize) -> Range<usize>

Returns the range that the column j occupies in self.row_indices().

Safety

The behavior is undefined if j >= self.ncols().

impl<I: Index> SymbolicSparseColMat<I>

pub fn try_new_from_indices( nrows: usize, ncols: usize, indices: &[(I, I)], ) -> Result<(Self, ValuesOrder<I>), CreationError>

Create a new symbolic structure, and the corresponding order for the numerical values from pairs of indices (row, col).

pub fn try_new_from_nonnegative_indices( nrows: usize, ncols: usize, indices: &[(I::Signed, I::Signed)], ) -> Result<(Self, ValuesOrder<I>), CreationError>

Create a new symbolic structure, and the corresponding order for the numerical values from pairs of indices (row, col).

Negative indices are ignored.