faer::sparse

Type Alias SparseRowMatMut

pub type SparseRowMatMut<'a, I, T, Rows = usize, Cols = usize> = SparseRowMat<Mut<'a, I, T, Rows, Cols>>;

view over sparse row major matrix

see SymbolicSparseRowMatRef

Aliased Type

#[repr(transparent)]
pub struct SparseRowMatMut<'a, I, T, Rows = usize, Cols = usize>(pub Mut<'a, I, T, Rows, Cols>);

Tuple Fields
0: Mut<'a, I, T, Rows, Cols>
Implementations

impl<'a, Rows: Shape, Cols: Shape, I: Index, T> SparseRowMatMut<'a, I, T, Rows, Cols>

pub fn new( symbolic: SymbolicSparseRowMatRef<'a, I, Rows, Cols>, val: &'a mut [T], ) -> Self

see SparseRowMatRef::new

pub fn parts(self) -> (SymbolicSparseRowMatRef<'a, I, Rows, Cols>, &'a [T])

see SparseRowMatRef::parts

pub fn parts_mut( self, ) -> (SymbolicSparseRowMatRef<'a, I, Rows, Cols>, &'a mut [T])

see SparseRowMatRef::parts

pub fn symbolic(&self) -> SymbolicSparseRowMatRef<'a, I, Rows, Cols>

see SparseRowMatRef::symbolic

pub fn val(self) -> &'a [T]

see SparseRowMatRef::val

pub fn val_mut(self) -> &'a mut [T]

see SparseRowMatRef::val

pub fn val_of_row(self, i: Idx<Rows>) -> &'a [T]

see SparseRowMatRef::val_of_row

pub fn val_of_row_mut(self, j: Idx<Rows>) -> &'a mut [T]

see SparseRowMatRef::val_of_row

pub fn as_shape<V: Shape, H: Shape>( self, nrows: V, ncols: H, ) -> SparseRowMatRef<'a, I, T, V, H>

see SparseRowMatRef::as_shape

pub fn as_shape_mut<V: Shape, H: Shape>( self, nrows: V, ncols: H, ) -> SparseRowMatMut<'a, I, T, V, H>

see SparseRowMatRef::as_shape

pub fn get(self, row: Idx<Rows>, col: Idx<Cols>) -> Option<&'a T>

see SparseRowMatRef::get

pub fn get_mut(self, row: Idx<Rows>, col: Idx<Cols>) -> Option<&'a mut T>

returns a reference to the value at the given index, or None if the symbolic structure doesn’t contain it, or contains multiple indices with the given index

panics

panics if row >= self.nrows()

panics if col >= self.ncols() see SparseRowMatRef::get

pub fn as_dyn(self) -> SparseRowMatRef<'a, I, T>

see SparseRowMatRef::as_dyn

pub fn as_dyn_mut(self) -> SparseRowMatMut<'a, I, T>

see SparseRowMatRef::as_dyn

pub fn transpose(self) -> SparseColMatRef<'a, I, T, Cols, Rows>

see SparseRowMatRef::transpose

pub fn transpose_mut(self) -> SparseColMatMut<'a, I, T, Cols, Rows>

see SparseRowMatRef::transpose

pub fn conjugate(self) -> SparseRowMatRef<'a, I, T::Conj, Rows, Cols>
where T: Conjugate,

see SparseRowMatRef::conjugate

pub fn conjugate_mut(self) -> SparseRowMatMut<'a, I, T::Conj, Rows, Cols>
where T: Conjugate,

see SparseRowMatRef::conjugate

pub fn adjoint(self) -> SparseColMatRef<'a, I, T::Conj, Cols, Rows>
where T: Conjugate,

see SparseRowMatRef::adjoint

pub fn adjoint_mut(self) -> SparseColMatMut<'a, I, T::Conj, Cols, Rows>
where T: Conjugate,

see SparseRowMatRef::adjoint

pub fn canonical(self) -> SparseRowMatRef<'a, I, T::Canonical, Rows, Cols>
where T: Conjugate,

see SparseRowMatRef::canonical

pub fn canonical_mut(self) -> SparseRowMatMut<'a, I, T::Canonical, Rows, Cols>
where T: Conjugate,

see SparseRowMatRef::canonical

pub fn to_col_major( &self, ) -> Result<SparseColMat<I, T::Canonical, Rows, Cols>, FaerError>
where T: Conjugate,

see SparseRowMatRef::to_col_major

pub fn to_dense(&self) -> Mat<T::Canonical, Rows, Cols>
where T: Conjugate,

see SparseRowMatRef::to_dense

pub fn triplet_iter( self, ) -> impl 'a + Iterator<Item = Triplet<Idx<Rows>, Idx<Cols>, &'a T>>
where Rows: 'a, Cols: 'a,

see SparseRowMatRef::triplet_iter