faer::sparse

Type Alias SymbolicSparseRowMat

pub type SymbolicSparseRowMat<I, Rows = usize, Cols = usize> = SymbolicSparseRowMat<Own<I, Rows, Cols>>;

owning symbolic structure of sparse matrix in row format, either compressed or uncompressed

see SymbolicSparseRowMatRef

Aliased Type

struct SymbolicSparseRowMat<I, Rows = usize, Cols = usize>(pub Own<I, Rows, Cols>);

Fields
0: Own<I, Rows, Cols>
Implementations

impl<Rows: Shape, Cols: Shape, I: Index> SymbolicSparseRowMat<I, Rows, Cols>

pub unsafe fn new_unchecked( nrows: Rows, ncols: Cols, row_ptr: Vec<I>, row_nnz: Option<Vec<I>>, col_idx: Vec<I>, ) -> Self

see SymbolicSparseRowMatRef::new_unchecked

pub fn new_checked( nrows: Rows, ncols: Cols, row_ptr: Vec<I>, row_nnz: Option<Vec<I>>, col_idx: Vec<I>, ) -> Self

see SymbolicSparseRowMatRef::new_checked

pub fn new_unsorted_checked( nrows: Rows, ncols: Cols, row_ptr: Vec<I>, row_nnz: Option<Vec<I>>, col_idx: Vec<I>, ) -> Self

see SymbolicSparseRowMatRef::new_unsorted_checked

pub fn parts(&self) -> (Rows, Cols, &[I], Option<&[I]>, &[I])

see SymbolicSparseRowMatRef::parts

pub fn into_parts(self) -> (Rows, Cols, Vec<I>, Option<Vec<I>>, Vec<I>)

see SymbolicSparseRowMatRef::parts

pub fn nrows(&self) -> Rows

see SymbolicSparseRowMatRef::nrows

pub fn ncols(&self) -> Cols

see SymbolicSparseRowMatRef::ncols

pub fn shape(&self) -> (Rows, Cols)

see SymbolicSparseRowMatRef::as_shape

pub fn transpose(&self) -> SymbolicSparseColMatRef<'_, I, Cols, Rows>

see SymbolicSparseRowMatRef::transpose

pub fn into_transpose(self) -> SymbolicSparseColMat<I, Cols, Rows>

see SymbolicSparseRowMatRef::transpose

pub fn to_owned(&self) -> Result<SymbolicSparseRowMat<I, Rows, Cols>, FaerError>

see SymbolicSparseRowMatRef::to_owned

pub fn to_col_major( &self, ) -> Result<SymbolicSparseColMat<I, Rows, Cols>, FaerError>

see SymbolicSparseRowMatRef::to_col_major

pub fn compute_nnz(&self) -> usize

see SymbolicSparseRowMatRef::compute_nnz

pub fn row_ptr(&self) -> &[I]

see SymbolicSparseRowMatRef::row_ptr

pub fn row_nnz(&self) -> Option<&[I]>

see SymbolicSparseRowMatRef::row_nnz

pub fn col_idx(&self) -> &[I]

see SymbolicSparseRowMatRef::col_idx

pub fn row_range(&self, i: Idx<Rows>) -> Range<usize>

see SymbolicSparseRowMatRef::row_range

pub unsafe fn row_range_unchecked(&self, i: Idx<Rows>) -> Range<usize>

see SymbolicSparseRowMatRef::row_range_unchecked

pub fn col_idx_of_row_raw(&self, i: Idx<Rows>) -> &[Idx<Cols, I>]

see SymbolicSparseRowMatRef::col_idx_of_row_raw

pub fn col_idx_of_row( &self, i: Idx<Rows>, ) -> impl '_ + Clone + ExactSizeIterator + DoubleEndedIterator<Item = Idx<Cols>>

see SymbolicSparseRowMatRef::col_idx_of_row

pub fn as_shape<V: Shape, H: Shape>( &self, nrows: V, ncols: H, ) -> SymbolicSparseRowMatRef<'_, I, V, H>

see SymbolicSparseRowMatRef::as_shape

pub fn into_shape<V: Shape, H: Shape>( self, nrows: V, ncols: H, ) -> SymbolicSparseRowMat<I, V, H>

see SymbolicSparseRowMatRef::as_shape

pub fn as_dyn(&self) -> SymbolicSparseRowMatRef<'_, I>

see SymbolicSparseRowMatRef::as_dyn

pub fn into_dyn(self) -> SymbolicSparseRowMat<I>

see SymbolicSparseRowMatRef::as_dyn

pub fn as_ref(&self) -> SymbolicSparseRowMatRef<'_, I, Rows, Cols>

Returns a view over the symbolic structure of self.

pub fn try_new_from_indices( nrows: Rows, ncols: Cols, idx: &[Pair<Idx<Rows, I>, Idx<Cols, I>>], ) -> Result<(Self, Argsort<I>), CreationError>

create a new symbolic structure, and the corresponding order for the numerical values from pairs of indices

pub fn try_new_from_nonnegative_indices( nrows: Rows, ncols: Cols, idx: &[Pair<MaybeIdx<Rows, I>, MaybeIdx<Cols, I>>], ) -> Result<(Self, Argsort<I>), CreationError>

create a new symbolic structure, and the corresponding order for the numerical values from pairs of indices

negative indices are ignored