Module faer::stats

Statistics-related utilities.

Structs

Normal

The normal distribution, N(mean, std_dev**2).

NormalCol

The normal distribution, N(mean, std_dev**2) for 0 <= i < nrows.

NormalMat

The normal distribution, N(mean, std_dev**2) for 0 <= i < nrows, 0 <= j < ncols.

NormalRow

The normal distribution, N(mean, std_dev**2) for 0 <= j < ncols.

StandardCol

The standard distribution. Samples uniformly distributed values for 0 <= i < nrows.

StandardMat

The standard distribution. Samples uniformly distributed values for 0 <= i < nrows, 0 <= j < ncols.

StandardNormalCol

The standard normal distribution, N(0, 1) for 0 <= i < nrows.

StandardNormalMat

The standard normal distribution, N(0, 1) for 0 <= i < nrows, 0 <= j < ncols.

StandardNormalRow

The standard normal distribution, N(0, 1) for 0 <= j < ncols.

StandardRow

The standard distribution. Samples uniformly distributed values for 0 <= j < ncols.

UnitaryMat

Uniformly samples a unitary matrix from the unitary group, in the sense of the Haar measure.

Enums

NanHandling

Specifies how missing values should be handled in mean and variance computations.

Functions

col_mean

Computes the mean of the columns of mat and stores the result in out.

col_varm

Computes the variance of the columns of mat given their mean, and stores the result in out.

row_mean

Computes the mean of the rows of mat and stores the result in out.

row_varm

Computes the variance of the rows of mat given their mean, and stores the result in out.