faer::stats

Module prelude

Re-exports

pub use super::CwiseColDistribution;

pub use super::CwiseMatDistribution;

pub use super::CwiseRowDistribution;

pub use super::DistributionExt;

pub use super::UnitaryMat;

Structs

ComplexDistribution

A generic random value distribution for complex numbers.

Standard

A generic random value distribution, implemented for many primitive types. Usually generates values with a numerically uniform distribution, and with a range appropriate to the type.

StandardNormal

Samples floating-point numbers according to the normal distribution N(0, 1) (a.k.a. a standard normal, or Gaussian). This is equivalent to Normal::new(0.0, 1.0) but faster.

StdRng

The standard RNG. The PRNG algorithm in StdRng is chosen to be efficient on the current platform, to be statistically strong and unpredictable (meaning a cryptographically secure PRNG).

ThreadRng

A reference to the thread-local generator

Traits

CryptoRng

A marker trait used to indicate that an RngCore or BlockRngCore implementation is supposed to be cryptographically secure.

Distribution

Types (distributions) that can be used to create a random instance of T.

IteratorRandom

Extension trait on iterators, providing random sampling methods.

Rng

An automatically-implemented extension trait on RngCore providing high-level generic methods for sampling values and other convenience methods.

RngCore

The core of a random number generator.

SeedableRng

A random number generator that can be explicitly seeded.

SliceRandom

Extension trait on slices, providing random mutation and sampling methods.

Functions

random

Generates a random value using the thread-local random number generator.

thread_rng

Retrieve the lazily-initialized thread-local random number generator, seeded by the system. Intended to be used in method chaining style, e.g. thread_rng().gen::<i32>(), or cached locally, e.g. let mut rng = thread_rng();. Invoked by the Default trait, making ThreadRng::default() equivalent.