faer

Struct Scale

#[repr(transparent)]
pub struct Scale<T>(pub T);

scaling factor for multiplying matrices.

Tuple Fields
0: T
Implementations

impl<T> Scale<T>

pub fn from_ref(value: &T) -> &Self

create a reference to a scaling factor from a reference to a value.

pub fn from_mut(value: &mut T) -> &mut Self

create a mutable reference to a scaling factor from a mutable reference to a value.