faer

Trait Auto

pub trait Auto<T> {
    // Required method
    fn auto() -> Self;
}

like Default, but with an extra type parameter so that algorithm hyperparameters can be tuned per scalar type.

Required Methods

fn auto() -> Self

returns the default value for the type T

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.