faer

Trait ShapeIdx

pub trait ShapeIdx {
    type Idx<I: Index>: Unbind<I> + Ord + Eq;
    type IdxInc<I: Index>: Unbind<I> + Ord + Eq + From<Idx<Self, I>>;
    type MaybeIdx<I: Index>: Unbind<I::Signed> + Ord + Eq;
}

base trait for Shape

Required Associated Types

type Idx<I: Index>: Unbind<I> + Ord + Eq

type that can be used to index into a range

type IdxInc<I: Index>: Unbind<I> + Ord + Eq + From<Idx<Self, I>>

type that can be used to partition a range

type MaybeIdx<I: Index>: Unbind<I::Signed> + Ord + Eq

either an index or a negative value

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.