faer

Trait Unbind

pub trait Unbind<I = usize>:
    Send
    + Sync
    + Copy
    + Debug
    + Seal {
    // Required methods
    unsafe fn new_unbound(idx: I) -> Self;
    fn unbound(self) -> I;
}

sealed trait for types that can be created from “unbound” values, as long as their struct preconditions are upheld

Required Methods

unsafe fn new_unbound(idx: I) -> Self

creates new value

safety

safety invariants must be upheld

fn unbound(self) -> I

returns the unbound value, unconstrained by safety invariants

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.