faer::utils

Module bound

compile-time bound-checked indexing types

Structs

Array

array of length equal to the value tied to 'n.

Dim

lifetime branded length

Idx

lifetime branded index.

IdxInc

lifetime branded partition index.

IdxIncOne

index equal to zero ro one

MaybeIdx

i value smaller than the size corresponding to the lifetime 'n, or none.

MaybeIdxOne

index equal to zero ro one, or a sentinel value

One

dimension equal to one

Partition

splits a range into two segments.

Zero

index equal to zero