faer::utils::bound

Struct Dim

pub struct Dim<'n> { /* private fields */ }

lifetime branded length

safety

the type’s safety invariant is that all instances of this type with the same lifetime correspond to the same length.

Implementations

impl<'n> Dim<'n>

pub fn with<R>(dim: usize, f: impl for<'dim> FnOnce(Dim<'dim>) -> R) -> R

create new branded value with the value dim.

pub const unsafe fn new_unbound(dim: usize) -> Self

create new branded value with an arbitrary brand.

safety

see struct safety invariant.

pub fn new(dim: usize, guard: Guard<'n>) -> Self

create new branded value with a unique brand.

pub const fn unbound(self) -> usize

returns the unconstrained value.

pub const fn partition<'head, 'tail>( self, midpoint: IdxInc<'n>, head: Guard<'head>, tail: Guard<'tail>, ) -> Partition<'head, 'tail, 'n>

partitions self into two segments as specifiedd by the midpoint.

pub fn head_partition<'head, 'tail>( self, head: Dim<'head>, tail: Guard<'tail>, ) -> Partition<'head, 'tail, 'n>

partitions self into two segments.

pub fn advance(self, start: Idx<'n>, len: usize) -> IdxInc<'n>

returns start advanced by len units, saturated to self

pub fn indices( self, ) -> impl Clone + ExactSizeIterator + DoubleEndedIterator<Item = Idx<'n>>

returns an iterator over the indices between 0 and self.

pub fn par_indices(self) -> impl IndexedParallelIterator<Item = Idx<'n>>

returns an iterator over the indices between 0 and self.

impl<'size> Dim<'size>

pub fn check<I: Index>(self, idx: I) -> Idx<'size, I>

check that the index is bounded by self, or panic otherwise.

pub fn idx<I: Index>(self, idx: I) -> Idx<'size, I>

check that the index is bounded by self, or panic otherwise.

pub fn idx_inc<I: Index>(self, idx: I) -> IdxInc<'size, I>

check that the index is bounded by self, or panic otherwise.

pub fn try_check<I: Index>(self, idx: I) -> Option<Idx<'size, I>>

check that the index is bounded by self, or return none otherwise.