faer::utils::bound

Struct Idx

pub struct Idx<'n, I: Index = usize> { /* private fields */ }

lifetime branded index.

safety

the type’s safety invariant is that all instances of this type are valid indices for [Dim<'n>] and less than or equal to i::signed::max.

Implementations

impl<'n, I: Index> Idx<'n, I>

pub const unsafe fn new_unbound(idx: I) -> Self

create new branded value with an arbitrary brand.

safety

see struct safety invariant.

pub unsafe fn new_unchecked(idx: I, dim: Dim<'n>) -> Self

create new branded value with the same brand as Dim.

safety

the behavior is undefined unless idx < dim and idx <= i::signed::max.

pub fn new_checked(idx: I, dim: Dim<'n>) -> Self

create new branded value with the same brand as Dim.

panics

panics unless idx < dim and idx <= i::signed::max.

pub const fn unbound(self) -> I

returns the unconstrained value.

pub fn zx(self) -> Idx<'n>

zero-extends the internal value into a usize.

impl<'n> Idx<'n>

pub fn truncate<I: Index>(self) -> Idx<'n, I>

truncate self to a smaller type i.

impl<'n, I: Index> Idx<'n, I>

pub const fn to_incl(self) -> IdxInc<'n, I>

returns the index, bounded inclusively by the value tied to 'n.

pub fn next(self) -> IdxInc<'n, I>

returns the next index, bounded inclusively by the value tied to 'n.

pub fn excl(self) -> IdxInc<'n, I>

returns the index, bounded inclusively by the value tied to 'n.

pub fn from_slice_mut_checked<'a>( slice: &'a mut [I], size: Dim<'n>, ) -> &'a mut [Idx<'n, I>]

assert that the values of slice are all bounded by size.

pub unsafe fn from_slice_mut_unchecked<'a>( slice: &'a mut [I], ) -> &'a mut [Idx<'n, I>]

assume that the values of slice are all bounded by the value tied to 'n.

pub fn from_slice_ref_checked<'a>( slice: &'a [I], size: Dim<'n>, ) -> &'a [Idx<'n, I>]

assert that the values of slice are all bounded by size.

pub unsafe fn from_slice_ref_unchecked<'a>(slice: &'a [I]) -> &'a [Idx<'n, I>]

assume that the values of slice are all bounded by the value tied to 'n.