faer::utils::bound

Struct MaybeIdx

pub struct MaybeIdx<'n, I: Index = usize> { /* private fields */ }

i value smaller than the size corresponding to the lifetime 'n, or none.

Implementations

impl<'n, I: Index> MaybeIdx<'n, I>

pub fn from_index(idx: Idx<'n, I>) -> Self

returns an index value.

pub fn none() -> Self

returns a none value.

pub fn new_checked(idx: I::Signed, size: Dim<'n>) -> Self

returns a constrained index value if idx is nonnegative, none otherwise.

pub unsafe fn new_unchecked(idx: I::Signed, size: Dim<'n>) -> Self

returns a constrained index value if idx is nonnegative, none otherwise.

pub unsafe fn new_unbound(idx: I) -> Self

returns a constrained index value if idx is nonnegative, none otherwise.

pub fn unbound(self) -> I

returns the inner value.

pub fn idx(self) -> Option<Idx<'n, I>>

returns the index if available, or none otherwise.

pub fn sx(self) -> MaybeIdx<'n>

sign extend the value.

pub fn from_slice_mut_checked<'a>( slice: &'a mut [I::Signed], size: Dim<'n>, ) -> &'a mut [MaybeIdx<'n, I>]

assert that the values of slice are all bounded by size.

pub unsafe fn from_slice_mut_unchecked<'a>( slice: &'a mut [I::Signed], ) -> &'a mut [MaybeIdx<'n, I>]

assume that the values of slice are all bounded by the value tied to 'n.

pub fn from_slice_ref_checked<'a>( slice: &'a [I::Signed], size: Dim<'n>, ) -> &'a [MaybeIdx<'n, I>]

assert that the values of slice are all bounded by size.

pub fn as_slice_ref<'a>(slice: &'a [MaybeIdx<'n, I>]) -> &'a [I::Signed]

convert a constrained slice to an unconstrained one.

pub unsafe fn from_slice_ref_unchecked<'a>( slice: &'a [I::Signed], ) -> &'a [MaybeIdx<'n, I>]

assume that the values of slice are all bounded by the value tied to 'n.