Struct faer::utils::slice::SliceGroupMut

pub struct SliceGroupMut<'a, E: Entity, T: 'a = <E as Entity>::Unit>(/* private fields */);

Analogous to a mutable reference to a slice for groups.

Implementations

impl<'a, E: Entity, T> SliceGroupMut<'a, E, T>

pub fn new(slice: GroupFor<E, &'a mut [T]>) -> Self

Create a new SliceGroup from a group of mutable slice references.

pub fn into_inner(self) -> GroupFor<E, &'a mut [T]>

Consume self to return the internally stored group of mutable slice references.

pub fn as_arrays_mut<const N: usize>( self, ) -> (SliceGroupMut<'a, E, [T; N]>, SliceGroupMut<'a, E, T>)

Decompose self into a mutable slice of arrays of size N, and a remainder part with length < N.

impl<'a, E: Entity> SliceGroupMut<'a, E>

pub fn read(&self, idx: usize) -> E

Read the element at position idx.

pub unsafe fn read_unchecked(&self, idx: usize) -> E

Read the element at position idx, without bound checks.

Safety

The behavior is undefined if idx >= self.len().

pub fn write(&mut self, idx: usize, value: E)

Write value to the location at position idx.

pub unsafe fn write_unchecked(&mut self, idx: usize, value: E)

Write value to the location at position idx, without bound checks.

Safety

The behavior is undefined if idx >= self.len().

pub fn fill_zero(&mut self)

Fill the slice with zeros.

impl<'a, E: Entity, T> SliceGroupMut<'a, E, T>

pub fn get_mut(self, idx: usize) -> RefGroupMut<'a, E, T>

Get a RefGroupMut pointing to the element at position idx.

pub unsafe fn get_unchecked_mut(self, idx: usize) -> RefGroupMut<'a, E, T>

Get a RefGroupMut pointing to the element at position idx.

Safety

The behavior is undefined if idx >= self.len().

pub fn get(self, idx: usize) -> RefGroup<'a, E, T>

Get a RefGroup pointing to the element at position idx.

pub unsafe fn get_unchecked(self, idx: usize) -> RefGroup<'a, E, T>

Get a RefGroup pointing to the element at position idx, without bound checks.

Safety

The behavior is undefined if idx >= self.len().

pub fn is_empty(&self) -> bool

Checks whether the slice is empty.

pub fn len(&self) -> usize

Returns the length of the slice.

pub fn subslice(self, range: Range<usize>) -> Self

Returns the subslice of self from the start to the end of the provided range.

pub unsafe fn subslice_unchecked(self, range: Range<usize>) -> Self

Returns the subslice of self from the start to the end of the provided range, without bound checks.

Safety

The behavior is undefined if range.start > range.end or range.end > self.len().

pub fn into_mut_iter(self) -> impl Iterator<Item = RefGroupMut<'a, E, T>>

Returns an iterator of RefGroupMut over the elements of the slice.

pub fn split_at(self, idx: usize) -> (Self, Self)

Split self at the midpoint idx, and return the two parts.

pub fn into_chunks_exact( self, chunk_size: usize, ) -> (impl Iterator<Item = SliceGroupMut<'a, E, T>>, Self)

Returns an iterator of slices over chunks of size chunk_size, and the remainder of the slice.