faer

Module mat

rectangular matrix

Modules

generic

generic Mat wrapper

Structs

Mut

see super::MatMut

Own

see super::Mat

Ref

see super::MatRef

Traits

AsMat

trait for owning matrix types

AsMatMut

trait for types that can be converted to a matrix view

AsMatRef

trait for types that can be converted to a matrix view

MatIndex

represents a type that can be used to slice a matrix, such as an index or a range of indices

Type Aliases

Mat

heap allocated resizable matrix, similar to a 2d alloc::vec::Vec

MatMut

mutable view over a matrix, similar to a mutable reference to a 2d strided slice

MatRef

immutable view over a matrix, similar to an immutable reference to a 2d strided slice