faer

Module prelude

useful imports for general usage of the library

Re-exports

pub use super::Par;

pub use super::Scale;

pub use super::c32;

pub use super::c64;

pub use super::col;

pub use super::col;

pub use super::mat;

pub use super::mat;

pub use super::row;

pub use super::row;

pub use col::Col;

pub use col::ColMut;

pub use col::ColRef;

pub use mat::Mat;

pub use mat::MatMut;

pub use mat::MatRef;

pub use row::Row;

pub use row::RowMut;

pub use row::RowRef;

pub use super::linalg::solvers::DenseSolve;

pub use super::linalg::solvers::Solve;

pub use super::linalg::solvers::SolveLstsq;

pub use sparse::SparseColMat;

pub use sparse::SparseColMatMut;

pub use sparse::SparseColMatRef;

pub use sparse::SparseRowMat;

pub use sparse::SparseRowMatMut;

pub use sparse::SparseRowMatRef;

Macros

col

creates a col::Col containing the arguments

mat

creates a Mat containing the arguments.

row

creates a row::Row containing the arguments

unzip

used to undo the zipping by the zip! macro.

zip

zips together matrix of the same size, so that coefficient-wise operations can be performed on their elements.

Traits

IntoConst

Consume a mutable reference to produce an immutable one.

Reborrow

Immutable reborrowing.

ReborrowMut

Mutable reborrowing.

Functions

default

see Default