faer::sparse

Module linalg

sparse linear algebra module. contains low level routines and the implementation of their corresponding high level wrappers

Modules

amd

approximate minimum degree ordering.

cholesky

computes the Cholesky decomposition (either $LL^\top$, $LTL^\top$, or $LBL^\top$) of a given sparse matrix. see crate::linalg::cholesky for more info.

colamd

approximate minimum degree column ordering.

lu

computes the $LU$ decomposition of a given sparse matrix. see faer::linalg::lu for more info

matmul

sparse matrix multiplication

qr

computes the $QR$ decomposition of a given sparse matrix. see crate::linalg::qr for more info

solvers

high-level sparse matrix solvers

triangular_solve

sparse matrix triangular solve

Structs

SupernodalThreshold

nonnegative threshold controlling when the supernodal factorization is used

SymbolicSupernodalParams

tuning parameters for the supernodal factorizations

Enums

LltError

sparse $ll^\top$ error

LuError

sparse $lu$ error.