faer::sparse::linalg::qr

Module simplicial

simplicial factorization module

a simplicial factorization is one that processes the elements of the $QR$ factors of the input matrix by single elements, rather than by blocks. this is more efficient if the $QR$ factors are very sparse

Structs

SimplicialQrRef

$QR$ factors containing both the symbolic and numeric representations

SymbolicSimplicialQr

symbolic structure of the $QR$ decomposition

Functions

factorize_simplicial_numeric_qr_scratch

computes the size and alignment of the workspace required to compute the numerical $QR$ factorization of the matrix whose structure was used to produce the symbolic structure

factorize_simplicial_numeric_qr_unsorted

computes the numerical $QR$ factorization of $A$.

factorize_simplicial_symbolic_qr

computes the symbolic $QR$ factorization of a matrix $A$, given the outputs of the pre-factorization steps

factorize_simplicial_symbolic_qr_scratch

computes the size and alignment of the workspace required to compute the symbolic $QR$ factorization of a matrix with dimensions (nrows, ncols)