faer::sparse

Struct SymbolicSparseColMat

pub struct SymbolicSparseColMat<I, Rows = usize, Cols = usize> { /* private fields */ }

owning symbolic structure of sparse matrix in column format, either compressed or uncompressed

see SymbolicSparseColMatRef

Implementations

impl<Rows: Shape, Cols: Shape, I: Index> SymbolicSparseColMat<I, Rows, Cols>

pub unsafe fn new_unchecked( nrows: Rows, ncols: Cols, col_ptr: Vec<I>, col_nnz: Option<Vec<I>>, row_idx: Vec<I>, ) -> Self

creates a new symbolic matrix view without checking its invariants

safety

see type level documentation.

pub fn new_checked( nrows: Rows, ncols: Cols, col_ptr: Vec<I>, col_nnz: Option<Vec<I>>, row_idx: Vec<I>, ) -> Self

creates a new symbolic matrix view after checking its invariants

safety

see type level documentation.

pub fn new_unsorted_checked( nrows: Rows, ncols: Cols, col_ptr: Vec<I>, col_nnz: Option<Vec<I>>, row_idx: Vec<I>, ) -> Self

creates a new symbolic matrix view after checking its invariants (excluding soft invariants)

safety

see type level documentation.

pub fn parts(&self) -> (Rows, Cols, &[I], Option<&[I]>, &[I])

see SymbolicSparseColMatRef::parts

pub fn into_parts(self) -> (Rows, Cols, Vec<I>, Option<Vec<I>>, Vec<I>)

see SymbolicSparseColMatRef::parts

pub fn nrows(&self) -> Rows

returns the number of rows of the matrix

pub fn ncols(&self) -> Cols

returns the number of columns of the matrix

pub fn shape(&self) -> (Rows, Cols)

returns the number of rows and columns of the matrix

pub fn transpose(&self) -> SymbolicSparseRowMatRef<'_, I, Cols, Rows>

returns a view over the transpose of self

pub fn into_transpose(self) -> SymbolicSparseRowMat<I, Cols, Rows>

returns the transpose of self

pub fn to_owned(&self) -> Result<SymbolicSparseColMat<I, Rows, Cols>, FaerError>

see SymbolicSparseColMatRef::to_owned

pub fn to_row_major( &self, ) -> Result<SymbolicSparseRowMat<I, Rows, Cols>, FaerError>

see SymbolicSparseColMatRef::to_row_major

pub fn compute_nnz(&self) -> usize

see SymbolicSparseColMatRef::compute_nnz

pub fn col_ptr(&self) -> &[I]

see SymbolicSparseColMatRef::col_ptr

pub fn col_nnz(&self) -> Option<&[I]>

see SymbolicSparseColMatRef::col_nnz

pub fn row_idx(&self) -> &[I]

see SymbolicSparseColMatRef::row_idx

pub fn col_range(&self, j: Idx<Cols>) -> Range<usize>

see SymbolicSparseColMatRef::col_range

pub unsafe fn col_range_unchecked(&self, j: Idx<Cols>) -> Range<usize>

see SymbolicSparseColMatRef::col_range_unchecked

pub fn row_idx_of_col_raw(&self, j: Idx<Cols>) -> &[Idx<Rows, I>]

see SymbolicSparseColMatRef::row_idx_of_col_raw

pub fn row_idx_of_col( &self, j: Idx<Cols>, ) -> impl '_ + Clone + ExactSizeIterator + DoubleEndedIterator<Item = Idx<Rows>>

see SymbolicSparseColMatRef::row_idx_of_col

pub fn as_shape<V: Shape, H: Shape>( &self, nrows: V, ncols: H, ) -> SymbolicSparseColMatRef<'_, I, V, H>

see SymbolicSparseColMatRef::as_shape

pub fn into_shape<V: Shape, H: Shape>( self, nrows: V, ncols: H, ) -> SymbolicSparseColMat<I, V, H>

see SymbolicSparseColMatRef::as_shape

pub fn as_dyn(&self) -> SymbolicSparseColMatRef<'_, I>

see SymbolicSparseColMatRef::as_dyn

pub fn into_dyn(self) -> SymbolicSparseColMat<I>

see SymbolicSparseColMatRef::as_dyn

pub fn as_ref(&self) -> SymbolicSparseColMatRef<'_, I, Rows, Cols>

Returns a view over the symbolic structure of self.

pub fn try_new_from_indices( nrows: Rows, ncols: Cols, idx: &[Pair<Idx<Rows, I>, Idx<Cols, I>>], ) -> Result<(Self, Argsort<I>), CreationError>

create a new symbolic structure, and the corresponding order for the numerical values from pairs of indices

pub fn try_new_from_nonnegative_indices( nrows: Rows, ncols: Cols, idx: &[Pair<MaybeIdx<Rows, I>, MaybeIdx<Cols, I>>], ) -> Result<(Self, Argsort<I>), CreationError>

create a new symbolic structure, and the corresponding order for the numerical values from pairs of indices

negative indices are ignored