faer

Module stats

statistics and randomness functionality

Modules

prelude

Structs

CwiseColDistribution

CwiseMatDistribution

CwiseRowDistribution

UnitaryMat

Enums

NanHandling

Specifies how missing values should be handled in mean and variance computations.

Traits

DistributionExt

Functions

col_mean

computes the mean of the columns of mat and stores the result in out

col_varm

computes the variance of the columns of mat and stores the result in out

row_mean

computes the mean of the rows of mat and stores the result in out

row_varm

computes the variance of the rows of mat and stores the result in out